% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLCalc.R
\name{checkDoubleTimes}
\alias{checkDoubleTimes}
\title{checkDoubleTimes}
\usage{
checkDoubleTimes(dt.df, acceptablePercentage = 25)
}
\arguments{
\item{dt.df}{data frame with the doubling times that need to be checked for high variation}

\item{acceptablePercentage}{What percentage either up or down that is acceptable in the computed difference between doubling times. default = 25%}
}
\description{
Calculate the area under the curve as determined by the survival percentage for an experiment.
}
