% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{cmb}
\alias{cmb}
\title{For combining a parameter and data list in RTMB so a data object can be explicitly defined}
\usage{
cmb(f, d)
}
\arguments{
\item{f}{Parameter list}

\item{d}{Data list}
}
\description{
For combining a parameter and data list in RTMB so a data object can be explicitly defined
}
\examples{
\dontrun{
  obj <- RTMB::MakeADFun(cmb(sabie_RTMB, data), parameters = parameters, map = mapping, random = random, silent = TRUE)
}
}
