% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGA.R
\name{create.plots}
\alias{create.plots}
\title{create.plots}
\usage{
create.plots(locationofRAW = homedir)
}
\description{
This function generates plots from the run of OGA, it wants a path to the RAW
files generated from OGA
}
