% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{dbeta_symmetric}
\alias{dbeta_symmetric}
\title{Symmetric Beta Function}
\usage{
dbeta_symmetric(p_val, p_ub, p_lb, p_prsd, log = TRUE)
}
\arguments{
\item{p_val}{Parameter value}

\item{p_ub}{Upper Bound of Parameter}

\item{p_lb}{Lower Bound of Parameter}

\item{p_prsd}{SD of parameter, higher values have a stronger penalty on bounds, lower values have a more difuse penalty on bounds}

\item{log}{whether or not to return the log likelihood}
}
\value{
Returns likelihood values from a symmetric beta
}
\description{
Symmetric Beta Function
}
\keyword{internal}
