% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{ddirichlet}
\alias{ddirichlet}
\title{Dirichlet Likelihood}
\usage{
ddirichlet(x, alpha, log = TRUE)
}
\arguments{
\item{x}{vector of values}

\item{alpha}{Expected values w/ concentration sum(alpha)}

\item{log}{Whether to give log or not}
}
\value{
Returns likelihood values form a dirichlet
}
\description{
Dirichlet Likelihood
}
