% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{ddirmult}
\alias{ddirmult}
\title{Dirichlet Mutlinomial Likelihood
From https://github.com/James-Thorson/CCSRA/blob/main/inst/executables/CCSRA_v9.cpp}
\usage{
ddirmult(obs, pred, Ntotal, ln_theta, give_log = TRUE)
}
\arguments{
\item{obs}{Vector of observed values in proportions}

\item{pred}{Vector or predicted values in proportions}

\item{Ntotal}{Input sample size scalar}

\item{ln_theta}{Weighting parameter in log space}

\item{give_log}{Whether or not likelihood is in log space}
}
\value{
returns likelihood values from a dirihclet multinomial
}
\description{
Dirichlet Mutlinomial Likelihood
From https://github.com/James-Thorson/CCSRA/blob/main/inst/executables/CCSRA_v9.cpp
}
