% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{dnbinom_robust_noint}
\alias{dnbinom_robust_noint}
\title{Negative binomial that can take non-integer values}
\usage{
dnbinom_robust_noint(x, log_mu, log_var_minus_mu, give_log = TRUE)
}
\arguments{
\item{x}{observations}

\item{log_mu}{log mu}

\item{log_var_minus_mu}{log var minus mu - reparameterize negbin}

\item{give_log}{whether to give log}
}
\value{
Returns likelihood values from a robust negative binomial
}
\description{
Negative binomial that can take non-integer values
}
