% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_mcmc.R
\name{do_mcmc}
\alias{do_mcmc}
\title{Run MCMC using rtmbstan}
\usage{
do_mcmc(obj, chains, cores, iter, thin, bounds = Inf, laplace = FALSE, ...)
}
\arguments{
\item{obj}{Object built from ADFUN via RTMB}

\item{chains}{Number of chains to run}

\item{cores}{Number of cores to use}

\item{iter}{Number of MCMC iterations to run}

\item{thin}{Thinning rate}

\item{bounds}{Uniform bounds to constrain parameter bounds}

\item{laplace}{Whether or not to do laplace approximation for random effects and MCMC for fixed effects}

\item{...}{Additional arguments for tmbstan (e.g., adapt_delta)}
}
\value{
MCMC list object from rtmbstan
}
\description{
Run MCMC using rtmbstan
}
\examples{
\dontrun{
obj <- RTMB::MakeADFun(cmb(SPoRC_rtmb, data), parameters = parameters, map = mapping)
mcmc <- do_mcmc(obj = obj,
                chains = 4,
                cores = 4,
                iter = 10000,
                thin = 10,
                bounds = 1.5,
                laplace = FALSE,
                adapt_delta = 0.99)
saveRDS(mcmc, here('output', 'MCMC_Model_23.5.rds'))
}
}
