% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{dpois_noint}
\alias{dpois_noint}
\title{Poisson that can take non-integer values}
\usage{
dpois_noint(x, pred, give_log = TRUE)
}
\arguments{
\item{x}{observations}

\item{pred}{predicted}

\item{give_log}{if giving log likelihood}
}
\value{
Returns likelihood values from a poisson
}
\description{
Poisson that can take non-integer values
}
\keyword{internal}
