% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equilibrium_Pop_Iter.R
\name{Equilibrium_PopSim_Iter}
\alias{Equilibrium_PopSim_Iter}
\title{Get equilibrium age structure for simulation via n iterations}
\usage{
Equilibrium_PopSim_Iter(
  n_regions = n_regions,
  n_sexes = n_sexes,
  n_ages = n_ages,
  init_iter = init_iter,
  NAA,
  M,
  init_F,
  fish_sel,
  r0,
  rec_sexratio,
  movement_matrix
)
}
\arguments{
\item{n_regions}{Number of reginos}

\item{n_sexes}{Number of sexes}

\item{n_ages}{Number of ages}

\item{init_iter}{Number of initial iterations to do}

\item{NAA}{Initial numbers at age array dimensioned by year, region, age, sex, and 1}

\item{M}{Natural mortality array dimensioned by year, region, age, sex, and 1}

\item{init_F}{Initial fishing mortality array dimensioned by year, region, fleet, and 1}

\item{fish_sel}{Fishery selectivity array dimensioned by year, region, age, sex, fleet, and 1}

\item{r0}{virgin or mean recruitment array dimensioned by year, region, and 1}

\item{rec_sexratio}{Recruitment sexratio array dimensioned by year, region, sex, and 1}

\item{movement_matrix}{Movement array, dimensioned by region from, region to, year, age, sex, and 1}
}
\description{
Get equilibrium age structure for simulation via n iterations
}
\keyword{internal}
