% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Run RTMB model}
\usage{
fit_model(
  data,
  parameters,
  mapping,
  random = NULL,
  newton_loops = 3,
  silent = FALSE
)
}
\arguments{
\item{data}{Data list}

\item{parameters}{Parameter list}

\item{mapping}{Mapping list}

\item{random}{Character of random effects to integrate out}

\item{newton_loops}{Number of newton loops to run to get gradients down}

\item{silent}{Boolean on whether or not model run is silent}
}
\value{
Returns a list object that is optimized, with results outputted from the RTMB model
}
\description{
Run RTMB model
}
\examples{
\dontrun{
model <- fit_model(data,
                 parameters,
                 mapping,
                 random = NULL,
                 newton_loops = 3)
}
}
