% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_3d_precision.R
\name{Get_3d_precision}
\alias{Get_3d_precision}
\title{Constructor algorithin for correlations within ages, years, and cohort (Work in Progress)}
\usage{
Get_3d_precision(
  n_ages,
  n_yrs,
  pcorr_age,
  pcorr_year,
  pcorr_cohort,
  ln_var_value,
  Var_Type
)
}
\arguments{
\item{n_ages}{Number of ages}

\item{n_yrs}{Number of years}

\item{pcorr_age}{correlations for age}

\item{pcorr_year}{correaltions for year}

\item{pcorr_cohort}{correlaitons for cohort}

\item{ln_var_value}{log space variance}

\item{Var_Type}{variance type == 0, marginal (stationary and slower run time), == 1 conditional (non-statationary, faster run time)}
}
\value{
Sparse precision matrix dimensioned by n_ages * n_years, n_ages * n_years
}
\description{
Constructor algorithin for correlations within ages, years, and cohort (Work in Progress)
}
\keyword{internal}
