% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{get_AR1_CorrMat}
\alias{get_AR1_CorrMat}
\title{Function to fill in an n x n correlation AR(1) matrix}
\usage{
get_AR1_CorrMat(n, rho)
}
\arguments{
\item{n}{Number of bins}

\item{rho}{correaltion parameter}
}
\value{
correlation matrix for an ar1 process
}
\description{
Function to fill in an n x n correlation AR(1) matrix
}
\examples{
\dontrun{
mat <- get_AR1_CorrMat(10, 0.5)
}
}
\keyword{internal}
