% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{get_beta_scaled_pars}
\alias{get_beta_scaled_pars}
\title{Get scaled alpha beta parameters for a scaled beta distribution (for steepness) from https://stackoverflow.com/questions/75165770/beta-distribution-with-bounds-at-0-1-0-5}
\usage{
get_beta_scaled_pars(low, high, mu, sigma)
}
\arguments{
\item{low}{lower bound}

\item{high}{upper bound}

\item{mu}{mean in normal space}

\item{sigma}{sigma in normal space}
}
\value{
Returns beta distribution parmeters with bounds
}
\description{
Get scaled alpha beta parameters for a scaled beta distribution (for steepness) from https://stackoverflow.com/questions/75165770/beta-distribution-with-bounds-at-0-1-0-5
}
\keyword{internal}
