% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Comp_Likelihoods.R
\name{Get_Comp_Likelihoods}
\alias{Get_Comp_Likelihoods}
\title{Gives negative log liklelihood values for composition data for a given year and a given fleet (fishery or survey)}
\usage{
Get_Comp_Likelihoods(
  Exp,
  Obs,
  ISS,
  Wt_Mltnml,
  ln_theta_agg,
  ln_theta,
  LN_corr_pars = 0,
  LN_corr_pars_agg = 0,
  Comp_Type,
  Likelihood_Type,
  n_regions,
  n_bins,
  n_sexes,
  age_or_len,
  AgeingError,
  use,
  comp_agg_type
)
}
\arguments{
\item{Exp}{Expected values (catch at age or survey index at age) indexed for a given year and fleet (structured as a matrix by age and sex)}

\item{Obs}{Observed values (catch at age or survey index at age) indexed for a given year and fleet (structured as a matrix by age and sex)}

\item{ISS}{Input sample size indexed for a given year and fleet (structured as a vector w/ sexes)}

\item{Wt_Mltnml}{Mutlinomial weight (if any) for a given fleet (structured as a vector w/ sexes)}

\item{ln_theta_agg}{Log overdispersion parameter if comp_type == 0, but we want to estsimate either a dirichlet or multinomial}

\item{ln_theta}{Log theta overdispersion for Dirichlet mutlinomial (scalar or vector depending on if 'Split' or 'Joint')}

\item{LN_corr_pars}{Logistic normal correlation parameters (dimensioned by n_regions, n_sexes, and 3 parameters)}

\item{LN_corr_pars_agg}{Logistic normal correlation parameters if comps are aggregated (just dimensioned by length of 1 value)}

\item{Comp_Type}{Composition Parameterization Type (== 0, aggregated comps by sex, == 1, split comps by sex and region (no implicit sex and region ratio information),
== 2, joint comps across sexes but split by region (implicit sex ratio information, but not region information)), == 3, joint comps across sexes and regions (implicit sex and region ratio information)}

\item{Likelihood_Type}{Composition Likelihood Type (== 0, Multinomial, == 1 Dirichlet Multinomial)}

\item{n_regions}{number of regions modeled}

\item{n_bins}{number of bins modeled}

\item{n_sexes}{Number of sexes modeled}

\item{age_or_len}{Age or length comps (== 0, Age, == 1, Length)}

\item{AgeingError}{Ageing Error matrix}

\item{use}{Vector of 0s and 1s corresponding to regions (==0, don't have obs and dont' use, ==1, have obs and use)}

\item{comp_agg_type}{How to aggregate data (if aggregating)}
}
\value{
Returns negative log likelihood for composition data (age and/or length)
}
\description{
Gives negative log liklelihood values for composition data for a given year and a given fleet (fishery or survey)
}
