% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{get_Constant_CorrMat}
\alias{get_Constant_CorrMat}
\title{Constant correlation matrix}
\usage{
get_Constant_CorrMat(n, rho)
}
\arguments{
\item{n}{Number of bins}

\item{rho}{correaltion parameter}
}
\value{
constant correlation matrix
}
\description{
Constant correlation matrix
}
\examples{
\dontrun{
mat <- get_Constant_CorrMat(10, 0.5)
}
}
\keyword{internal}
