% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Plots.R
\name{get_data_fitted_plot}
\alias{get_data_fitted_plot}
\title{Title Get Data Fitted to Plot}
\usage{
get_data_fitted_plot(data)
}
\arguments{
\item{data}{data list fed into SPoCK}
}
\value{
A plot of data that were fitted to
}
\description{
Title Get Data Fitted to Plot
}
\examples{
\dontrun{
get_data_fitted_plot(data)
}
}
