% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Det_Recruitment.R
\name{Get_Det_Recruitment}
\alias{Get_Det_Recruitment}
\title{Get Deterministic Recruitment}
\usage{
Get_Det_Recruitment(
  recruitment_model,
  y,
  rec_lag,
  R0,
  h,
  n_ages,
  WAA,
  MatAA,
  natmort,
  SSB_vals
)
}
\arguments{
\item{recruitment_model}{== 0, mean recruitment, == 1 beverton holt recruitment with steepness}

\item{y}{year}

\item{rec_lag}{recruitment lag for indexing SSB year}

\item{R0}{virgin or mean recruitment}

\item{h}{steepness}

\item{n_ages}{number of ages}

\item{WAA}{vector of weight at age}

\item{MatAA}{vector of maturity at age}

\item{natmort}{vector of natural mortality at age}

\item{SSB_vals}{spawning stock biomass value}
}
\value{
Determinstic recruitment values from mean recruitment or beverton holt
}
\description{
Get Deterministic Recruitment
}
