% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_Population.R
\name{Get_Feedback_Data}
\alias{Get_Feedback_Data}
\title{Helper function that obtains data for a simulation with a feedback loop}
\usage{
Get_Feedback_Data(
  sim_env,
  sim_list,
  y,
  sim,
  skeleton_data,
  skeleton_parameters,
  skeleton_mapping
)
}
\arguments{
\item{sim_env}{Simulation environment defined using Simulate_Pop_Feedback}

\item{sim_list}{Simulation list that defines the operating model options}

\item{y}{y index used for annual cycle}

\item{sim}{sim index used for annual cycle}

\item{skeleton_data}{Skeleton data list for estimation method (SPoCK) where all dimensions are defined using the total number of simulation years}

\item{skeleton_parameters}{Skeleton parameters list for estimation method (SPoCK) where all dimensions are defined using the total number of simulation years}

\item{skeleton_mapping}{Skeleton mapping list for estimation method (SPoCK) where all dimensions are defined using the total number of simulation years}
}
\value{
Returns data, parameters, and mapping list truncated to match the feedback year
}
\description{
Helper function that obtains data for a simulation with a feedback loop
}
