% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Plots.R
\name{get_idx_fits_plot}
\alias{get_idx_fits_plot}
\title{Title Get Index Fits Plot}
\usage{
get_idx_fits_plot(data, rep)
}
\arguments{
\item{data}{data list from SPoCK}

\item{rep}{report list from SPoCK}
}
\value{
A plot of fitted values to various indices
}
\description{
Title Get Index Fits Plot
}
\examples{
\dontrun{
get_idx_fits_plot(data,rep)
}
}
