% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_mcmc.R
\name{get_mcmc_ssb_rec}
\alias{get_mcmc_ssb_rec}
\title{Get ssb and recruitment from MCMC using posterior samples}
\usage{
get_mcmc_ssb_rec(mcmc_obj, tmb_obj)
}
\arguments{
\item{mcmc_obj}{MCMC object from tmbstan}

\item{tmb_obj}{RTMB object from MakeADFUN}
}
\value{
returns a dataframe of recruitment and ssb posterior samples and a plot of ssb and recruitment with mean, lwr 95 quantile, and upr 95 quantile.
}
\description{
Get ssb and recruitment from MCMC using posterior samples
}
