% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_PE_loglik.R
\name{Get_move_PE_loglik}
\alias{Get_move_PE_loglik}
\title{Title Get Movement Process Error Likelihoods}
\usage{
Get_move_PE_loglik(
  PE_model,
  PE_pars,
  logit_devs,
  map_move_devs,
  do_recruits_move
)
}
\arguments{
\item{PE_model}{Process error model values}

\item{PE_pars}{Process error parameters}

\item{logit_devs}{Deviations}

\item{map_move_devs}{movement deviations to share}

\item{do_recruits_move}{Whether recruits move (0, don't move, 1 move)}
}
\value{
numeric value of log likelihood (in positive space)
}
\description{
Title Get Movement Process Error Likelihoods
}
\keyword{internal}
