% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Plots.R
\name{get_nLL_plot}
\alias{get_nLL_plot}
\title{Title Get plot of negative log likelihood values}
\usage{
get_nLL_plot(rep, sd_rep)
}
\arguments{
\item{rep}{Report list from SPoCK}

\item{sd_rep}{sdreport list from SPoCK}
}
\value{
Plot of negative log likelihood values
}
\description{
Title Get plot of negative log likelihood values
}
\examples{
\dontrun{
get_nLL_plot(rep, sd_rep)
}
}
