% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Plots_Tables.R
\name{get_selex_plot}
\alias{get_selex_plot}
\title{Get Fishery and Survey Selectivity Plots}
\usage{
get_selex_plot(rep, model_names, Selex_Type = "age")
}
\arguments{
\item{rep}{List of n_models of `SPoRC` report lists}

\item{model_names}{Vector of model names}

\item{Selex_Type}{Character vector specifying whether to output age or length-based selectivity (age, length)}
}
\value{
Plots of terminal year fishery and survey selectivity by fleet, region, and sex across models
}
\description{
Get Fishery and Survey Selectivity Plots
}
\examples{
\dontrun{
get_selex_plot(list(rep1, rep2), c("Model1", "Model2"))
}
}
