% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Tagging_Mortality.R
\name{Get_Tagging_Mortality}
\alias{Get_Tagging_Mortality}
\title{Get values for total tag mortality and tag fishing mortality}
\usage{
Get_Tagging_Mortality(
  tag_selex,
  tag_natmort,
  Fmort,
  natmort,
  Tag_Shed,
  fish_sel,
  n_regions,
  n_ages,
  n_sexes,
  n_fish_fleets,
  y,
  what
)
}
\arguments{
\item{tag_selex}{Tag selectivity options, == 0 (uniform, with F from fleet 1 (dominant fleet)), == 1 (sex-averaged selectivity, with F from fleet 1 (dominant fleet)), == 2 (sex-specific selectivity, with F from fleet 1 (dominant fleet)), 3 (uniform with F summed across fleets), 4 (sex averaged with F summed across fleets, weighted sum), 5 (sex-specific with F summed across fleets, weighted sum)}

\item{tag_natmort}{Tag natural mortality options == 0 (averaged across sexes and ages), == 1 (averaged across sexes, but unique for ages), == 2 (sex-specific, but averaged across ages), == 3 (sex-and age-specific)}

\item{Fmort}{Array of fishing mortality, dimensioned by n_region, n_years, n_fish_fleets}

\item{natmort}{Array of fishing mortality, dimensioned by n_region, n_years, n_ages, n_sexes}

\item{Tag_Shed}{Scalar chronic tag shedding rate}

\item{fish_sel}{Array of fishery selectivity, dimensioned by n_region, n_years, n_ages, n_sexes, n_fish_fleetss}

\item{n_regions}{Number of regions}

\item{n_ages}{Number of ages}

\item{n_sexes}{Number of sexes}

\item{n_fish_fleets}{Number of fishery fleets}

\item{y}{Year index}

\item{what}{Whether to return Z or F (total or fishing mortality)}
}
\value{
Z or F values from tagging specifications
}
\description{
Get values for total tag mortality and tag fishing mortality
}
\keyword{internal}
