% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Plots.R
\name{get_ts_plot}
\alias{get_ts_plot}
\title{Title Get Time Series Plots}
\usage{
get_ts_plot(rep, sd_rep)
}
\arguments{
\item{rep}{RTMB report list}

\item{sd_rep}{RTMB sdreport list}
}
\value{
Plots of spawning biomass, total biomass, recruitment, and fishing mortality time-series
}
\description{
Title Get Time Series Plots
}
\examples{
\dontrun{
  get_ts_plot(rep, sd_rep)
}
}
