% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Reference_Points.R
\name{global_BH_Fmsy}
\alias{global_BH_Fmsy}
\title{Title Get Global FMSY from a Beverton-Holt (Spatial)}
\usage{
global_BH_Fmsy(pars, data)
}
\arguments{
\item{pars}{Parameter List}

\item{data}{Data List}
}
\description{
Title Get Global FMSY from a Beverton-Holt (Spatial)
}
\examples{
\dontrun{
rep <- obj$report(obj$env$last.par.best) # need to have an RTMB object first
data_list <- list() # set up data list
# Extract out relevant elements
n_ages <- length(data$ages) # number of ages
n_years <- length(data$years) # number of years
data_list$F_fract_flt <- rep$Fmort[1,n_years,] / sum(rep$Fmort[1,n_years,]) # get fleet F fraction
data_list$fish_sel <- array(rep$fish_sel[1,n_years,,1,], dim = c(n_ages, data$n_fish_fleets))
data_list$natmort <- rep$natmort[1,1,,1]
data_list$t_spwn <- 0
data_list$WAA <- data$WAA[1,1,,1]
data_list$MatAA <- data$MatAA[1,1,,1]
data_list$h <- rep$h_trans
data_list$R0 <- rep$R0

par_list <- list()
par_list$log_Fmsy <- log(0.1)

obj <- RTMB::MakeADFun(cmb(single_region_BH_Fmsy, data_list), parameters = par_list, map = NULL, silent = TRUE)
obj$optim <- stats::nlminb(obj$par, obj$fn, obj$gr, control = list(iter.max = 1e6, eval.max = 1e6, rel.tol = 1e-15))
obj$rep <- obj$report(obj$env$last.par.best)
obj$sdrep <- sdreport(obj)
}
}
\keyword{internal}
