% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{has_dependencies}
\alias{has_dependencies}
\title{Check availability of dependency packages}
\usage{
has_dependencies()
}
\value{
A named list with logical values indicating if each dependency is available
}
\description{
This function returns the availability status of all optional dependencies
used by the SPoCK package.
}
