% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{install_spock_dependencies}
\alias{install_spock_dependencies}
\title{Install SPoCK dependencies}
\usage{
install_spock_dependencies(force = FALSE)
}
\arguments{
\item{force}{Logical, whether to force reinstallation of dependencies even if they are already installed}
}
\description{
This function installs the optional dependencies required for full
functionality of the SPoCK package.
}
