% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{is_package_available}
\alias{is_package_available}
\title{Title Helper function to check package availbility}
\usage{
is_package_available(pkg)
}
\arguments{
\item{pkg}{package name character}
}
\description{
Title Helper function to check package availbility
}
\keyword{internal}
