% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGA.R
\name{ladderCreate}
\alias{ladderCreate}
\title{ladderCreate}
\usage{
ladderCreate(laddername = "ladder.csv", repladder = 3)
}
\arguments{
\item{laddername}{Name of the csv file that contains your calibration ladder data.}

\item{repladder}{number of technical replicates in your calibration ladder}
}
\value{
Returns a calibration ladder for OGA() analysis.
}
\description{
Function to create calibration ladder used by OGA.
The output of this function is a 3rd order polynomial which is the best fit of the data. This function expects to find a CSV file called
ladder.csv (or user defined name) in the current working directory; this should
be a user generated file. See included Example file for format of ladder.csv
}
