% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_mcmc.R
\name{mcmc_diag}
\alias{mcmc_diag}
\title{Model diagnostics for MCMC}
\usage{
mcmc_diag(mcmc_obj, tmb_sdrep, plot = TRUE, pars = NULL)
}
\arguments{
\item{mcmc_obj}{MCMC object from tmbstan}

\item{tmb_sdrep}{SD report fomr TMB}

\item{plot}{Whether or not to plot outputs}

\item{pars}{Parameters you want to plot out}
}
\value{
A list of plots with traditional MCMC diagnostics
}
\description{
Model diagnostics for MCMC
}
\examples{
\dontrun{
diagnostic_plots <- mcmc_diag(mcmc_obj, tmb_sdrep, TRUE, pars = 'ln_M')
}
}
