% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_mcmc.R
\name{mcmc_par_names}
\alias{mcmc_par_names}
\title{Gets parameter names from mcmc object}
\usage{
mcmc_par_names(mcmc_obj)
}
\arguments{
\item{mcmc_obj}{MCMC object from rtmbstan}
}
\value{
Returns a character vector of parameter names
}
\description{
Gets parameter names from mcmc object
}
\examples{
\dontrun{
par_names <- mcmc_par_names(mcmc_obj)
}
}
