% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mlt_rg_sable_data}
\alias{mlt_rg_sable_data}
\title{Sablefish data for multi region case study}
\format{
A list with multiple components needed for the multi (5) region sablefish model
}
\source{
Cheng et al. 2025, In review; Panmictic Panacea? Demonstrating Good Practices for Developing Spatial Stock Assessments through Application to Alaska Sablefish (Anoplopoma fimbria)
}
\usage{
mlt_rg_sable_data
}
\description{
A dataset containing the necessary elements for the Alaska sablefish spatial case study.
}
\keyword{datasets}
