% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mlt_rg_sable_mod_results}
\alias{mlt_rg_sable_mod_results}
\title{Sablefish model results for multi region case study}
\format{
A list object containing parameters, report, sdreport, and elements from a MakeADFun object
}
\source{
Cheng et al. 2025, In review; Panmictic Panacea? Demonstrating Good Practices for Developing Spatial Stock Assessments through Application to Alaska Sablefish (Anoplopoma fimbria)
}
\usage{
mlt_rg_sable_mod_results
}
\description{
A dataset containing all model results the Alaska sablefish multi region case study.
}
\keyword{datasets}
