% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGA.R
\name{OGA}
\alias{OGA}
\title{OGA (Outgrowth Analysis)}
\usage{
OGA(
  autoinput = TRUE,
  nestdataframe = dfs,
  homedir = getwd(),
  ladder = TRUE,
  laddername = "ladder.csv",
  repladder = 3,
  blank = TRUE,
  blankname = "BLANK",
  create.plot = TRUE,
  stringencyfilt = 3,
  frequencyfilt = 0.05,
  measureInterval = 15,
  lowerlimitslope = 0.01,
  maxinflectionpoint = 1.3,
  LimitNoGrowth = 0.9,
  statsSurvival = TRUE,
  bacteria = 45,
  laglimit = 0.1
)
}
\arguments{
\item{autoinput}{BOOL parameter, if TRUE then automatically read in all .csv files for outgrowth
analysis. these csv files are autoinputed into DFS, otherwise OGA expect to find a value for dfs.}

\item{nestdataframe}{if autoinput is FALSE then you need to pass instead a nested dataframe to analyze
to this paramater.}

\item{homedir}{Home Directory, Working Directory containing RAW .csv files from plate reader in Example.csv format.}

\item{ladder}{BOOL parameter, if TRUE then OGA finds calibration ladder, and calibrates RAW data.}

\item{laddername}{Name of ladder.csv file (in same directory as RAW .csv files).}

\item{repladder}{Number of technical replicates in ladder (see example).}

\item{blank}{Do you have blanks (wells with no innoculum) in the csv files (TRUE/FALSE)}

\item{blankname}{Name used for blank wells, this name is case sensitive and must match exactly with file.}

\item{create.plot}{BOOL parameter, if TRUE then OGA passes RAW files to create.plots to produce plots? (TRUE/FALSE).}

\item{stringencyfilt}{What order to use for stringency of Butterworth Filter.}

\item{frequencyfilt}{What cuttoff to use for frequency of Butterworth Filter.}

\item{measureInterval}{Time between OD measurements in minutes.}

\item{lowerlimitslope}{Slope above which growth is considered exponential.}

\item{maxinflectionpoint}{Maximum OD above which the exponential growth phase has ended.}

\item{LimitNoGrowth}{What is the OD limit to signal no growth detected/insufficient information for complete growth curve.}

\item{bacteria}{Limit below which doubling-time is marked as bacterial contamination, if measuring bacterial growth set = 0.5.}

\item{laglimit}{Minimum OD above which the exponential growth phase has begun.}

\item{stats}{BOOL parameter, if TRUE the OGA passes results files to statsSurvival(TRUE/FALSE)}
}
\description{
OGA calculates doubling times by subtracting blank wells, smoothing high frequency noise by a
Butterworth filter, picking the second limit by defining the inflection point of exponential
growth as determined by the 1st derivative (diff) ul, and first limit by a heuristically
defined slope change the doubling time is then calculated between these points.
}
