% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Plots_Tables.R
\name{plot_all_basic}
\alias{plot_all_basic}
\title{Plotting function for all basic quantities}
\usage{
plot_all_basic(data, rep, sd_rep, model_names, out_path)
}
\arguments{
\item{data}{List of n_models of `SPoRC` data lists}

\item{rep}{List of n_models of `SPoRC` report lists}

\item{sd_rep}{List of n_models of sd report lists from `SPoRC`}

\item{model_names}{Character vector of model names}

\item{out_path}{Path to the output directory. Users only need to specify the path.}
}
\value{
A series of plots compared across models outputted as a pdf in the specified directory
}
\description{
Plotting function for all basic quantities
}
\examples{
\dontrun{
plot_all_basic(
  data = list(data1, data2),
  rep = list(rep1, rep2),
  sd_rep = list(sd_rep1, sd_rep2),
  model_names = c("Model1", "Model2"),
  out_path = here::here()
)
}
}
