% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Model_Fits.R
\name{plot_resids}
\alias{plot_resids}
\title{Plots OSA residuals from outputs from get_osa. Much of this code is taken from the afscOM package, but with modificaitons to plot features.}
\usage{
plot_resids(osa_results)
}
\arguments{
\item{osa_results}{List object obtained from get_osa, that contains a dataframe of residuals and aggregated fits.}
}
\value{
A vareity of plots for OSA residuals (list)
}
\description{
Plots OSA residuals from outputs from get_osa. Much of this code is taken from the afscOM package, but with modificaitons to plot features.
}
\examples{
\dontrun{
comp_props <- get_comp_prop(data = data, rep = rep, age_labels = 2:31, len_labels = seq(41, 99, 2), year_labels = 1960:2024)
osa_results <- get_osa(obs_mat = comp_props$Obs_FishLen_mat,
                       exp_mat = comp_props$Pred_FishLen_mat,
                       N = 20 * data$Wt_FishAgeComps[1,1,1],
                       years = 1999:2023,
                       fleet = 1,
                       bins = 2:31,
                       comp_type = 1,
                       bin_label = "Age")

osa_plot <- plot_resids(osa_results)
}
}
