% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{post_optim_sanity_checks}
\alias{post_optim_sanity_checks}
\title{Post Optimization Model Convergence Checks}
\usage{
post_optim_sanity_checks(
  sd_rep,
  rep,
  gradient_tol = 0.001,
  se_tol = 100,
  corr_tol = 0.99
)
}
\arguments{
\item{sd_rep}{sd report list from a `SPoCK` model}

\item{rep}{report list from a `SPoCK` model}

\item{gradient_tol}{Value for maximum gradient tolerance to use}

\item{se_tol}{Value for maximum standard error tolerance to use}

\item{corr_tol}{Value for maximum correlation tolerance to use}
}
\description{
Post Optimization Model Convergence Checks
}
