% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_Compositions.R
\name{rdirM}
\alias{rdirM}
\title{Title Simulate dirichlet multinomial draws}
\usage{
rdirM(n, N, alpha)
}
\arguments{
\item{n}{Number of sims}

\item{N}{Sum of observations}

\item{alpha}{Concentration parameter}
}
\description{
Title Simulate dirichlet multinomial draws
}
\keyword{internal}
