% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_RV.R
\name{rinvgauss_rec}
\alias{rinvgauss_rec}
\title{Title Generate Recruitment Values based on Inverse Gaussian Distribution}
\usage{
rinvgauss_rec(sims, recruitment)
}
\arguments{
\item{sims}{Number of Simulations}

\item{recruitment}{Recruitment vector}
}
\value{
Random variables following an inverse gaussian
}
\description{
Title Generate Recruitment Values based on Inverse Gaussian Distribution
}
\keyword{internal}
