% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Dimensions.R
\name{Setup_Mod_Dim}
\alias{Setup_Mod_Dim}
\title{Set up model dimensions}
\usage{
Setup_Mod_Dim(
  years,
  ages,
  lens,
  n_regions,
  n_sexes,
  n_fish_fleets,
  n_srv_fleets,
  verbose = FALSE
)
}
\arguments{
\item{years}{Numeric vector of years.}

\item{ages}{Numeric vector of age classes.}

\item{lens}{Numeric vector of length bins; can be set to \code{1} if length data are not modeled.}

\item{n_regions}{Integer specifying the number of spatial regions.}

\item{n_sexes}{Integer specifying the number of sexes.}

\item{n_fish_fleets}{Integer specifying the number of fishery fleets.}

\item{n_srv_fleets}{Integer specifying the number of survey fleets.}

\item{verbose}{Logical flag indicating whether to print progress messages (default \code{FALSE}).}
}
\value{
A list containing three named elements:
\describe{
  \item{\code{data}}{List of data inputs dimensioned by the model dimensions.}
  \item{\code{parameters}}{List of model parameters initialized according to dimensions.}
  \item{\code{map}}{List of parameter mappings for model fitting.}
}
}
\description{
Set up model dimensions
}
