% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Movement.R
\name{Setup_Mod_Movement}
\alias{Setup_Mod_Movement}
\title{Setup model movement processes}
\usage{
Setup_Mod_Movement(
  input_list,
  do_recruits_move = 0,
  use_fixed_movement = 0,
  Fixed_Movement = NA,
  Use_Movement_Prior = 0,
  Movement_prior = NULL,
  Movement_ageblk_spec = NULL,
  Movement_yearblk_spec = NULL,
  Movement_sexblk_spec = NULL,
  cont_vary_movement = "none",
  Movement_cont_age_spec = "none",
  Movement_cont_year_spec = "none",
  Movement_cont_pe_pars_spec = "none",
  ...
)
}
\arguments{
\item{input_list}{List containing a data list, parameter list, and map list}

\item{do_recruits_move}{Whether recruits move, 1 (they move), 0 (they don't move), default is 0}

\item{use_fixed_movement}{Whether to use a fixed movement matrix, == 0 don't use, == 1 use}

\item{Fixed_Movement}{Input for a fixed movement matrix dimensioned by n_regions, n_regions, n_years, n_ages, n_sexes}

\item{Use_Movement_Prior}{Whether or not to use movement priors == 0 don't use, == 1 use}

\item{Movement_prior}{Movement prior values, can be dimension as a vector of (results in a constant prior across all regions from, years, ages, and sexes), or as an array by n_regions, n_regions, n_years, n_ages, n_sexes}

\item{Movement_ageblk_spec}{Either a character string specifiying "constant" age movement, or a ist object specifying which ages to group together and block (i.e., which ages to share parameters for). Each element in the list should be a vector specifying the range of ages for which to share or block parameters for. For example, list(c(1:6), c(7:10), c(11:n_ages)) has 3 age blocks, where ages 1-6 have the same paraemters, 7:10 have the same parameters (but diffeerent from 1:6), and 11:n_ages haave the same parameters (but different from 1:6 and 7:10).
The default behavior is to estimate unique movement parmaeters for all ages. To estimate age-invariant movement this list would be specified as: list(c(1:n_ages)).}

\item{Movement_yearblk_spec}{Either a character string specifiying "constant" year movement, list object specifying which years to group together and block (i.e., which years to share parameters for). Each element in the list should be a vector specifying the range of years for which to share or block parameters for. For example, list(c(1:6), c(7:10), c(11:n_years)) has 3 year blocks, where years 1-6 have the same paraemters, 7:10 have the same parameters (but diffeerent from 1:6), and 11:n_years haave the same parameters (but different from 1:6 and 7:10).
The default behavior is to estimate unique movement parameters for all years To estimate time-invariant movement this list would be specified as: list(c(1:n_years)).}

\item{Movement_sexblk_spec}{Either a character string specifiying "constant" sex movement, list object specifying which sexes to group together and block (i.e., which sexes to share parameters for). Each element in the list should be a vector specifying the range of sexes for which to share or block parameters for. For example, list(1, 2) has 2 sex blocks, where each sex has unique parameters
The default behavior is to estimate unique movement parameters for all sexes. To estimate sex-invariant movement this list would be specified as: list(c(1:n_sexes)).}

\item{cont_vary_movement}{Continuous varying movement 0 (not continuous), 1 (continuous)}

\item{Movement_cont_age_spec}{Continuous age deviations as random effects. A character specified at est_all or none (i.e., no age deviations uses none, vs all ages uses est_all)}

\item{Movement_cont_year_spec}{Continuous year deviations as random effects. A character specified at est_all or none (i.e., no year deviations uses none, vs all years uses est_all)}

\item{Movement_cont_pe_pars_spec}{Specification of process error parameters. Character string where options are est_shared_r, est_shared_a, est_shared_r_a, est_all, fix, none}

\item{...}{Additional inputs for starting values for movement parameters (move_pars)}
}
\description{
Setup model movement processes
}
