% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Observation_Processes.R
\name{Setup_Mod_Weighting}
\alias{Setup_Mod_Weighting}
\title{Set up SPoCK model weighting}
\usage{
Setup_Mod_Weighting(
  input_list,
  sablefish_ADMB,
  likelihoods,
  Wt_Catch = 1,
  Wt_FishIdx = 1,
  Wt_SrvIdx = 1,
  Wt_Rec = 1,
  Wt_F = 1,
  Wt_Tagging = 1,
  Wt_FishAgeComps,
  Wt_SrvAgeComps,
  Wt_FishLenComps,
  Wt_SrvLenComps
)
}
\arguments{
\item{input_list}{List containing a data list, parameter list, and map list}

\item{sablefish_ADMB}{Numeric indicating whether to mimic calculations for the sablefish ADMB model (1) or not (0)}

\item{likelihoods}{Numeric indicating whether to use ADMB likelihoods (0) or TMB likelihoods (1)}

\item{Wt_Catch}{Weight (lambda) applied to the overall catch dataset}

\item{Wt_FishIdx}{Weight (lambda) applied to the overall fishery index dataset}

\item{Wt_SrvIdx}{Weight (lambda) applied to the overall survey index dataset}

\item{Wt_Rec}{Weight (lambda) applied to recruitment penalty}

\item{Wt_F}{Weight (lambda) applied to fishing mortality penalty}

\item{Wt_Tagging}{Weight (lambda) applied to tagging data}

\item{Wt_FishAgeComps}{Weight (lambda) applied to fishery age compositions}

\item{Wt_SrvAgeComps}{Weight (lambda) applied to survey age compositions}

\item{Wt_FishLenComps}{Weight (lambda) applied to fishery length compositions}

\item{Wt_SrvLenComps}{Weight (lambda) applied to survey length compositions}
}
\description{
Set up SPoCK model weighting
}
