% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Biologicals.R
\name{Setup_Sim_Biologicals}
\alias{Setup_Sim_Biologicals}
\title{Set up simulation stuff for biologicals}
\usage{
Setup_Sim_Biologicals(
  n_sims = n_sims,
  n_yrs = n_yrs,
  n_regions = n_regions,
  n_ages = n_ages,
  n_sexes = n_sexes,
  base_M_value,
  M_pattern,
  base_WAA_values,
  WAA_pattern,
  base_Maturity_AA_values,
  Maturity_AA_pattern
)
}
\arguments{
\item{n_sims}{Number of simulations}

\item{n_yrs}{Number of years}

\item{n_regions}{Number of regions}

\item{n_ages}{Number of ages}

\item{n_sexes}{Number of sexes}

\item{base_M_value}{Base natural mortality value dimensioned by regions, ages, sexes}

\item{M_pattern}{Natural mortality pattern. Options include: constant}

\item{base_WAA_values}{Base weight-at-age values dimensioned by regions, ages, sexes}

\item{WAA_pattern}{Weight-at-age pattern. Options include: constant}

\item{base_Maturity_AA_values}{Base maturity values dimensioned by regions, ages, sexes}

\item{Maturity_AA_pattern}{Maturity pattern. Options include: constant}
}
\description{
Set up simulation stuff for biologicals
}
