% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Dimensions.R
\name{Setup_Sim_Dim}
\alias{Setup_Sim_Dim}
\title{Set up simulation dimensions}
\usage{
Setup_Sim_Dim(
  n_sims,
  n_yrs,
  n_regions,
  n_ages,
  n_sexes,
  n_fish_fleets,
  n_srv_fleets,
  run_feedback,
  feedback_start_yr
)
}
\arguments{
\item{n_sims}{Number of simulations}

\item{n_yrs}{Number of years}

\item{n_regions}{Number of regions}

\item{n_ages}{Number of ages}

\item{n_sexes}{Number of sexes}

\item{n_fish_fleets}{Number of fishery fleets}

\item{n_srv_fleets}{Number of survey fleets}

\item{run_feedback}{Boolean for whether to run or not run feedback management loop}

\item{feedback_start_yr}{If MSE is run, when is the first year feedback starts}
}
\description{
Set up simulation dimensions
}
