% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_Population.R
\name{Setup_sim_env}
\alias{Setup_sim_env}
\title{Constructs simulation objects in a new simulation environment for use in simulation functions}
\usage{
Setup_sim_env(sim_list)
}
\arguments{
\item{sim_list}{Simulation list objects}
}
\value{
A new simulation environment with objects from sim_list
}
\description{
Constructs simulation objects in a new simulation environment for use in simulation functions
}
\examples{
\dontrun{
sim_env <- Setup_sim_env(sim_list)
}
}
