% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Fishery.R
\name{Setup_Sim_FishMort}
\alias{Setup_Sim_FishMort}
\title{Setup values and dimensions of fishing mortality}
\usage{
Setup_Sim_FishMort(
  n_sims = n_sims,
  n_yrs = n_yrs,
  n_regions = n_regions,
  n_fish_fleets = n_fish_fleets,
  sigmaC,
  init_F_vals,
  Fmort_pattern,
  Fmort_start,
  Fmort_fct,
  proc_error,
  proc_error_sd
)
}
\arguments{
\item{n_sims}{Number of simulations}

\item{n_yrs}{Number of years}

\item{n_regions}{Number of regions}

\item{n_fish_fleets}{Number of fleets}

\item{sigmaC}{Observation error for catch}

\item{init_F_vals}{Initial F values dimensioend by region and fleet}

\item{Fmort_pattern}{Fishing mortality pattern as a matrix dimensioned by region and fleet (constant, linear, one-way, two-way)}

\item{Fmort_start}{Fishing mortality start values as a matrix dimensioned by region and fleet}

\item{Fmort_fct}{Fishing mortality values for factor increases or decreases as a matrix dimensioned by region and fleet}

\item{proc_error}{If we want to add process error to fishing mortality}

\item{proc_error_sd}{value of logrnomal sd for process error to fishing mortality}
}
\description{
Setup values and dimensions of fishing mortality
}
