% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Observation_Processes.R
\name{Setup_Sim_Observation_Proc}
\alias{Setup_Sim_Observation_Proc}
\title{Set up simulation observaiton processes}
\usage{
Setup_Sim_Observation_Proc(
  n_fish_fleets,
  n_srv_fleets,
  Comp_Structure,
  Comp_Srv_Like,
  Srv_Like_Pars,
  Comp_Fish_Like,
  Fish_Like_Pars,
  Tag_Like,
  Tag_Like_Pars
)
}
\arguments{
\item{n_fish_fleets}{Number of fishery fleets}

\item{n_srv_fleets}{Number of survey fleets}

\item{Comp_Structure}{Composition Structure (SpltSex_SpltRegion, JntSex_SpltRegion, JntSex_JntRegion)}

\item{Comp_Srv_Like}{Survey Composition Likelihoods (Multinomial, Dirichlet-Multinomial, iid-Logistic-Normal)}

\item{Srv_Like_Pars}{Parameters for Survey Composition Likelihoods}

\item{Comp_Fish_Like}{Fishery Composition Likelihoods (Multinomial, Dirichlet-Multinomial, iid-Logistic-Normal)}

\item{Fish_Like_Pars}{Parameters for Fishery Composition Likelihoods}

\item{Tag_Like}{Tag Likelihoods (Poisson, NegBin, Multinomial_Release, Multinomial_Recapture)}

\item{Tag_Like_Pars}{Tag Likelihood Parameters}
}
\description{
Set up simulation observaiton processes
}
