% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Rec.R
\name{Setup_Sim_Rec}
\alias{Setup_Sim_Rec}
\title{Set up recruitment dynamics for simulation}
\usage{
Setup_Sim_Rec(
  do_recruits_move,
  base_rec_sexratio,
  rec_sexratio_vary,
  base_r0,
  r0_vary,
  base_h,
  init_sigmaR,
  sigmaR,
  recruitment_opt,
  rec_dd,
  init_dd,
  sim_list,
  rec_lag
)
}
\arguments{
\item{do_recruits_move}{whether recruits move. Character string either specified as 'dont_move', or 'move'}

\item{base_rec_sexratio}{base recruitment sex-ratio value}

\item{rec_sexratio_vary}{whether recruitment sex-ratio varies. Options include: constant}

\item{base_r0}{base R0 or mean recruitment value}

\item{r0_vary}{whether r0 or mean recruitment varies. Options include: constant}

\item{base_h}{base steepness value}

\item{init_sigmaR}{Sigma R for initial devs}

\item{sigmaR}{Sigma R for everything else}

\item{recruitment_opt}{character string as either "mean_rec" or "bh_rec"}

\item{rec_dd}{recruitment density dependence; character string as either "global" or "local"}

\item{init_dd}{initial age density dependence; character string as either "global" or "local"}

\item{sim_list}{Simulation list}

\item{rec_lag}{Numeric, recruitment lag value}
}
\description{
Set up recruitment dynamics for simulation
}
