% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sgl_rg_ebswp_mod_results}
\alias{sgl_rg_ebswp_mod_results}
\title{EBS Walleye Pollock model results for single region case study}
\format{
A list object containing parameters, report, and elements from a MakeADFun object
}
\source{
2024 Federal EBS Walleye Pollock Assessment
}
\usage{
sgl_rg_ebswp_mod_results
}
\description{
A dataset containing all model results the EBS Walleye Pollock single region case study.
}
\keyword{datasets}
