% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sgl_rg_sable_data}
\alias{sgl_rg_sable_data}
\title{Sablefish data for single region case study}
\format{
A list with multiple components needed for the single region sablefish model
}
\source{
2024 Federal Alaska Sablefish Assessment
}
\usage{
sgl_rg_sable_data
}
\description{
A dataset containing the necessary elements for the Alaska sablefish case study.
}
\keyword{datasets}
