% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sgl_rg_sable_mod_results}
\alias{sgl_rg_sable_mod_results}
\title{Sablefish model results for single region case study}
\format{
A list object containing parameters, report, and elements from a MakeADFun object
}
\source{
2024 Federal Alaska Sablefish Assessment
}
\usage{
sgl_rg_sable_mod_results
}
\description{
A dataset containing all model results the Alaska sablefish single region case study.
}
\keyword{datasets}
