% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sgl_rg_sable_rep}
\alias{sgl_rg_sable_rep}
\title{Sablefish report for single region case study}
\format{
Report file from the single region sablefish case study
}
\source{
2024 Federal Alaska Sablefish Assessment
}
\usage{
sgl_rg_sable_rep
}
\description{
A report containing results for the Alaska sablefish case study.
}
\keyword{datasets}
