% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPoRC_rtmb.R
\name{SPoRC_rtmb}
\alias{SPoRC_rtmb}
\title{Generalized RTMB model}
\usage{
SPoRC_rtmb(pars, data)
}
\arguments{
\item{pars}{Parameter List}

\item{data}{Data List}
}
\description{
Generalized RTMB model
}
\keyword{internal}
