% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{stats}
\alias{stats}
\title{stats}
\usage{
stats(locationofresults = "Results", LimitNoGrowth = 0.9, rmflagged = TRUE)
}
\arguments{
\item{locationofresults}{path to the results of analysis.}

\item{LimitNoGrowth}{Parameter specifying the cutoff value for determining growth.}
}
\description{
stats takes raw results files and calculates the mean, SD and SEM of all wells with the same name.
}
