% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{statsSurvival}
\alias{statsSurvival}
\title{statsCLS}
\usage{
statsSurvival(
  locationofresults = "Results",
  LimitNoGrowth = 0.9,
  rmflagged = TRUE
)
}
\arguments{
\item{locationofresults}{path to the results of OGA analysis.}

\item{LimitNoGrowth}{Parameter specifying the cutoff value for determining growth. default=0.9}

\item{rmflagged}{This determines if flagged wells are removed from statistics. default=TRUE}
}
\description{
stats takes raw results files and calculates the mean, SD and SEM of all wells with the same name.
}
