% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLCalc.R
\name{SurvivalCalc}
\alias{SurvivalCalc}
\title{SurvivalCalc}
\usage{
SurvivalCalc(
  firstDay = 1,
  resultspath = getwd(),
  rmflagged = TRUE,
  stats = TRUE,
  measureInterval = 15
)
}
\arguments{
\item{firstDay}{Specify the first day of measurements, default=1}

\item{resultspath}{System path where the results are stored, default=current directory}

\item{stats}{Should the code run stats on the results (TRUE/FALSE), default=TRUE}

\item{measureInterval}{Interval in minutes for each measurement default=15}
}
\description{
SurvivalCalc does something cool.
}
