% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLCalc.R
\name{SurvivalPercentage}
\alias{SurvivalPercentage}
\title{SurvivalPercentage}
\usage{
SurvivalPercentage(RAWpath = getwd(), firstDay = 1, measureInterval = 15)
}
\arguments{
\item{RAWpath}{path where the raw data files are located. Default = current working directory}

\item{firstDay}{First day of measurements. Default = 1}

\item{measureInterval}{Time between each OD measurement in minutes. Default = 15}
}
\description{
This function computes percentage of cells still alive. This function looks for files
that have the filename of the syntax <experiment>_Day_###.csv where <experiment> can be
any unique ID to define the experiment, _Day_ is a marker to identify when the day is
going to be specified and ### is the actual day of measuring.
}
