% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_retrospective.R
\name{truncate_yr}
\alias{truncate_yr}
\title{Helper function to truncate data years, parameters, and mapping to conduct retrospective diagnostics. Called within do_retrospective function.}
\usage{
truncate_yr(j, data, parameters, mapping)
}
\arguments{
\item{j}{The years to truncate from the terminal year}

\item{data}{Data list used for the RTMB model}

\item{parameters}{Parameter list used for the RTMB model}

\item{mapping}{Mapping list used for the RTMB model}
}
\value{
List of data, parameters, and mapping that have truncated dimensions from the original data, parameters, and mapping list
}
\description{
Helper function to truncate data years, parameters, and mapping to conduct retrospective diagnostics. Called within do_retrospective function.
}
\examples{
\dontrun{
retro_list <- retro_truncate_year(j = 0, data, parameters, mapping) # does not remove any data
retro_list <- retro_truncate_year(j = 1, data, parameters, mapping) # removes last year of data
}
}
\keyword{internal}
