% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{collect_message}
\alias{collect_message}
\title{Title Helper function to collect messages}
\usage{
collect_message(...)
}
\arguments{
\item{...}{character vector of messages}
}
\description{
Title Helper function to collect messages
}
\keyword{internal}
