% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGA.R
\name{create.plots}
\alias{create.plots}
\title{create.plots}
\usage{
create.plots(locationofRAW = getwd())
}
\arguments{
\item{locationofRAW}{path to the RAW data files generated by OGA().}
}
\description{
This function generates plots from OGA results anaylsis, it wants a path to the RAW
files generated from OGA
}
