% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{dlogistnormal}
\alias{dlogistnormal}
\title{Logistic Normal Likelihood}
\usage{
dlogistnormal(obs, pred, Sigma, give_log = TRUE)
}
\arguments{
\item{obs}{Vector of observed values in numbers (can be integers or non-integers - vector length of n_bins)}

\item{pred}{Vector of predicted values in proportions (vector length of n_bins)}

\item{Sigma}{Covariance matrix}

\item{give_log}{whether or not to use log space likelihood}
}
\value{
Returns likelihood values from a logistic normal
}
\description{
Logistic Normal Likelihood
}
\keyword{internal}
