% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Do_Population_Projection.R
\name{Do_Population_Projection}
\alias{Do_Population_Projection}
\title{Title Do Population Projections}
\usage{
Do_Population_Projection(
  n_proj_yrs = 2,
  n_regions,
  n_ages,
  n_sexes,
  sexratio,
  n_fish_fleets,
  do_recruits_move = 0,
  recruitment,
  terminal_NAA,
  terminal_F,
  natmort,
  WAA,
  MatAA,
  fish_sel,
  Movement,
  f_ref_pt = NULL,
  b_ref_pt = NULL,
  HCR_function = NULL,
  recruitment_opt = "inv_gauss",
  fmort_opt = "HCR",
  F_input = NULL,
  t_spawn
)
}
\arguments{
\item{n_proj_yrs}{Number of projection years}

\item{n_regions}{Number of regions}

\item{n_ages}{Number of ages}

\item{n_sexes}{Number of sexes}

\item{sexratio}{Recruitment sex ratio}

\item{n_fish_fleets}{Number of fishery fleets}

\item{do_recruits_move}{Whether recruits move (0 == don't move, 1 == move)}

\item{recruitment}{Recruitment matrix dimensioned by n_regions, and n_yrs that we want to summarize across, or condition our projection on}

\item{terminal_NAA}{Terminal Numbers at Age dimensioned by n_regions, n_ages, n_sexes}

\item{terminal_F}{Terminal fishing mortality rate, dimensioned by n_regions, n_fish_fleets}

\item{natmort}{Natural mortality, dimensioned by n_regions, n_proj_yrs, n_ages, n_sexes}

\item{WAA}{Weight at age, dimensioned by n_regions, n_proj_yrs, n_ages, n_sexes}

\item{MatAA}{Maturity at age, dimensioned by n_regions, n_proj_yrs, n_ages, n_sexes}

\item{fish_sel}{Fishery selectivity, dimensioned by n_regions, n_proj_yrs, n_ages, n_sexes, n_fish_fleets}

\item{Movement}{Movement, dimensioned by n_regions, n_regions, n_proj_yrs, n_ages, n_sexes}

\item{f_ref_pt}{Fishing mortality reference point dimensioned by n_regions}

\item{b_ref_pt}{Biological reference point dimensioned by n_regions}

\item{HCR_function}{Function describing a harvest control rule. The function should always have the following arguments: x, which represents SSB, frp, which takes inputs of fishery reference points, and brp, which takes inputs of biological reference points. Any additional arguments should be specified with defaults or hard coded / fixed within the function. e}

\item{recruitment_opt}{Recruitment simulation option, where options are "inv_gauss", which simulates future recruitment based on the the recruitment values supplied using an inverse gaussian distribution, or "mean_rec", which takes the mean of the recruitment values supplied for a given region}

\item{fmort_opt}{Fishing Mortality option, which includes "HCR", which modifies the F reference point using a user supplied HCR_function, or "Input", which uses projected F values supplied by the user.}

\item{F_input}{Projected fishing mortality input, dimensioned by n_regions, n_proj_yrs}

\item{t_spawn}{Fraction time of spawning used to compute projected SSB}
}
\value{
A list containing projected F, catch, SSB, and Numbers at Age. (Objects are generally dimensioned in the following order: n_regions, n_yrs, n_ages, n_sexes, n_fleets)
}
\description{
Title Do Population Projections
}
\examples{
\dontrun{
# Setup necessary inputs
n_sims <- 1000
t_spawn <- 0
sexratio <- 0.5
n_proj_yrs <- 15
n_regions <- 1
n_ages <- length(data$ages)
n_sexes <- data$n_sexes
n_fish_fleets <- 2
do_recruits_move <- 0
terminal_NAA <- array(sabie_rtmb_model$rep$NAA[,length(data$years),,], dim = c(n_regions, n_ages, n_sexes))
WAA <- array(rep(data$WAA[,length(data$years),,], each = n_proj_yrs), dim = c(n_regions, n_proj_yrs, n_ages, n_sexes)) # weight at age
MatAA <- array(rep(data$MatAA[,length(data$years),,], each = n_proj_yrs), dim = c(n_regions, n_proj_yrs, n_ages, n_sexes)) # maturity at age
fish_sel <- array(rep(sabie_rtmb_model$rep$fish_sel[,length(data$years),,,], each = n_proj_yrs), dim = c(n_regions, n_proj_yrs, n_ages, n_sexes, n_fish_fleets)) # selectivity
Movement <- array(rep(sabie_rtmb_model$rep$Movement[,,length(data$years),,], each = n_proj_yrs), dim = c(n_regions, n_regions, n_proj_yrs, n_ages, n_sexes))
terminal_F <- array(sabie_rtmb_model$rep$Fmort[,length(data$years),], dim = c(n_regions, n_fish_fleets))
natmort <- array(sabie_rtmb_model$rep$natmort[,length(data$years),,], dim = c(n_regions, n_proj_yrs, n_ages, n_sexes))
recruitment <- array(sabie_rtmb_model$rep$Rec[,20:(length(data$years) - 2)], dim = c(n_regions, length(20:length(data$years))))

# Define reference points
spr_35 <- Get_Reference_Points(data = data,
                               rep = rep,
                               SPR_x = 0.35, t_spwn = 0, sex_ratio_f = 0.5,
                               calc_rec_st_yr = 20, rec_age = 2)

spr_40 <- Get_Reference_Points(data = data,
                               rep = rep,
                               SPR_x = 0.4, t_spwn = 0, sex_ratio_f = 0.5,
                               calc_rec_st_yr = 20, rec_age = 2)

spr_60 <- Get_Reference_Points(data = data,
                               rep = rep,
                               SPR_x = 0.6, t_spwn = 0, sex_ratio_f = 0.5,
                               calc_rec_st_yr = 20, rec_age = 2)

# Extract reference points
b40 <- spr_40$b_ref_pt
b60 <- spr_60$b_ref_pt
f40 <- spr_40$f_ref_pt
f35 <- spr_35$f_ref_pt
f60 <- spr_60$f_ref_pt
# Define the F inputs for each scenario (Based on BSAI Intro Report)
 F_inputs <- list(
   array(f40, dim = c(n_regions, n_proj_yrs)), # Scenario 1
   array(f40 * (f40 / 0.086), dim = c(n_regions, n_proj_yrs)), # Scenario 2
   array(mean(rowSums(sabie_rtmb_model$rep$Fmort[1, 61:65, ])), dim = c(n_regions, n_proj_yrs)),  # Scenario 3
   array(f60, dim = c(n_regions, n_proj_yrs)), # Scenario 4
   array(0, dim = c(n_regions, n_proj_yrs)), # Scenario 5
   array(f35, dim = c(n_regions, n_proj_yrs)), # Scenario 6
   array(c(rep(f40, 2), rep(f35, n_proj_yrs - 2)), dim = c(n_regions, n_proj_yrs)) # Scenario 7
 )

 # store SSB outputs
 all_scenarios_ssb <- array(0, dim = c(n_regions, n_proj_yrs, n_sims, length(F_inputs)))
 all_scenarios_catch <- array(0, dim = c(n_regions, n_proj_yrs, n_fish_fleets, n_sims, length(F_inputs)))

 for (i in seq_along(F_inputs)) {
   for (sim in 1:n_sims) {

     # do population projection
     out <- Do_Population_Projection(n_proj_yrs = n_proj_yrs,
                                     n_regions = n_regions,
                                     n_ages = n_ages,
                                     n_sexes = n_sexes,
                                     sexratio = sexratio,
                                     n_fish_fleets = n_fish_fleets,
                                     do_recruits_move = do_recruits_move,
                                     recruitment = recruitment,
                                     terminal_NAA = terminal_NAA,
                                     terminal_F = terminal_F,
                                     natmort = natmort,
                                     WAA = WAA,
                                     MatAA = MatAA,
                                     fish_sel = fish_sel,
                                     Movement = Movement,
                                     f_ref_pt = f_ref_pt,
                                     b_ref_pt = b_ref_pt,
                                     HCR_function = HCR_function,
                                     recruitment_opt = "inv_gauss",
                                     fmort_opt = 'Input',
                                     F_input = F_inputs[[i]],
                                     t_spawn = t_spawn
     )

     all_scenarios_ssb[,,sim,i] <- out$proj_SSB
     all_scenarios_catch[,,,sim,i] <- out$proj_Catch


   } # end sim loop
   print(i)
 } # end i loop
}
}
