% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Plots.R
\name{get_biological_plot}
\alias{get_biological_plot}
\title{Title Get Plots of Biological Quantities}
\usage{
get_biological_plot(data, rep)
}
\arguments{
\item{data}{data list used for SPoCK}

\item{rep}{report list used for SPoCK}
}
\value{
A list of plots for movement, natural mortality, weight-at-age, and maturity at age
}
\description{
Title Get Plots of Biological Quantities
}
\examples{
\dontrun{
get_biological_plot(data, rep)
}
}
