% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Plots.R
\name{get_data_fitted_plot}
\alias{get_data_fitted_plot}
\title{Get Data Fitted to Plot}
\usage{
get_data_fitted_plot(data, model_names)
}
\arguments{
\item{data}{List of n_models of `SPoCK` data lists}

\item{model_names}{Character vector of model names}
}
\value{
A plot of data that were fitted to across models
}
\description{
Get Data Fitted to Plot
}
\examples{
\dontrun{
get_data_fitted_plot(list(data1, data2), c("Model1", "Model2"))
}
}
