% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Det_Recruitment.R
\name{Get_Det_Recruitment}
\alias{Get_Det_Recruitment}
\title{Get Deterministic Recruitment}
\usage{
Get_Det_Recruitment(
  recruitment_model,
  recruitment_dd,
  y,
  rec_lag,
  R0,
  Rec_Prop,
  h,
  n_regions,
  n_ages,
  WAA,
  MatAA,
  natmort,
  SSB_vals
)
}
\arguments{
\item{recruitment_model}{== 0, mean recruitment, == 1 beverton holt recruitment with steepness}

\item{recruitment_dd}{Recruitment density dependence (0 == local, 1 == global)}

\item{y}{year}

\item{rec_lag}{recruitment lag for indexing SSB year}

\item{R0}{virgin or mean recruitment (global)}

\item{Rec_Prop}{Recruitment proportions to allocate global R0 if local density dependence}

\item{h}{Vector of steepness values by n_regions}

\item{n_regions}{Number of regions}

\item{n_ages}{number of ages}

\item{WAA}{Weight at age by region and age}

\item{MatAA}{Maturity by region and age}

\item{natmort}{Natural moratliaty by region and age}

\item{SSB_vals}{SSB matrix by regiona and year}
}
\value{
Vector of n_regions of deterministic recruitment values from mean recruitment or beverton holt
}
\description{
Get Deterministic Recruitment
}
\keyword{internal}
