% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Plots.R
\name{get_idx_fits_plot}
\alias{get_idx_fits_plot}
\title{Get Index Fits Plot}
\usage{
get_idx_fits_plot(data, rep, model_names)
}
\arguments{
\item{data}{List of n_models of `SPoCK` data lists}

\item{rep}{List of n_models of `SPoCK` report lists}

\item{model_names}{Vector of model names}
}
\value{
A plot of fitted values to various indices across models
}
\description{
Get Index Fits Plot
}
\examples{
\dontrun{
get_idx_fits_plot(list(data1, data2), list(rep1, rep2), c("Model1", "Model2"))
}
}
