% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Reference_Points.R
\name{Get_Reference_Points}
\alias{Get_Reference_Points}
\title{Wrapper function to get reference points}
\usage{
Get_Reference_Points(
  data,
  rep,
  SPR_x = NULL,
  t_spwn = 0,
  sex_ratio_f = 0.5,
  calc_rec_st_yr = 1,
  rec_age = 1,
  type = "single_region",
  what = "SPR"
)
}
\arguments{
\item{data}{Data list from RTMB}

\item{rep}{Report list from RTMB}

\item{SPR_x}{SPR percentage to target}

\item{t_spwn}{specified mortality time up until spawning}

\item{sex_ratio_f}{Sex ratio for females used to compute biological reference points}

\item{calc_rec_st_yr}{The first year in which mean recruitment is computed for}

\item{rec_age}{Actual recruitment age value}

\item{type}{Whether this is a "single_region" reference point or "multi_region"}

\item{what}{What kind of reference point "SPR" and "SPR_independent" (used for multi region, where SPR is calcualted for each region independently)}
}
\value{
A list object of fishing and biological reference points
}
\description{
Wrapper function to get reference points
}
\examples{
\dontrun{
f_40 <- Get_Reference_Points(data = data,
rep = rep,
SPR_x = 0.4,
t_spwn = 0,
type = "single_region",
what = "SPR")
f_40$F_x # estimated reference point
}
}
