% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Plots.R
\name{get_retrospective_plot}
\alias{get_retrospective_plot}
\title{Get Retrospective Plot}
\usage{
get_retrospective_plot(retro_output)
}
\arguments{
\item{retro_output}{Dataframe generated from do_retrospective}
}
\value{
A retrospective plot of recruitment and SSB
}
\description{
Get Retrospective Plot
}
\examples{
\dontrun{
# do retrospective
retro <- do_retrospective(n_retro = 7, # number of retro peels to run
data = data, # rtmb data
parameters = parameters, # rtmb parameters
mapping = mapping, # rtmb mapping
random = NULL, # if random effects are used
do_par = TRUE, # whether or not to parralleize
n_cores = 7, # if parallel, number of cores to use
do_francis = F, # if we want tod o Francis
n_francis_iter = NULL # Number of francis iterations to do
)
get_retrospective_plot(retro)
}
}
