% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Selex.R
\name{Get_Selex}
\alias{Get_Selex}
\title{Get Selectivity}
\usage{
Get_Selex(
  Selex_Model,
  TimeVary_Model,
  ln_Pars,
  ln_seldevs,
  Region,
  Year,
  Age,
  Sex
)
}
\arguments{
\item{Selex_Model}{integer for selectivity model}

\item{TimeVary_Model}{Time-varying selectivity model == 0, constant or blocked, == 1 iid, == 2 random walk, == 3 3dar1 marginal, == 4 3dar1 conditional, ==5 2dar1}

\item{ln_Pars}{log selectivity parameters}

\item{ln_seldevs}{selectivity deviations (either in array or matrix and needs to match the time_vary model specification)}

\item{Region}{Region index value}

\item{Year}{Year index value}

\item{Age}{vector of Ages}

\item{Sex}{Sex index value}
}
\description{
Get Selectivity
}
\keyword{internal}
