% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Plots.R
\name{get_selex_plot}
\alias{get_selex_plot}
\title{Title Get Fishery and Survey Selectivity Plots}
\usage{
get_selex_plot(rep)
}
\arguments{
\item{rep}{RTMB report list}
}
\value{
Plots of fishery and survey selectivity by fleet, region, and sex
}
\description{
Title Get Fishery and Survey Selectivity Plots
}
\examples{
\dontrun{
get_selex_plot(rep)
}
}
