% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Plots_Tables.R
\name{get_ts_plot}
\alias{get_ts_plot}
\title{Get Time Series Plots}
\usage{
get_ts_plot(rep, sd_rep, model_names)
}
\arguments{
\item{rep}{List of n_models of `SPoRC` report lists}

\item{sd_rep}{List of n_models of `SPoRC` sdreport lists}

\item{model_names}{Vector of model names}
}
\value{
Plots of spawning biomass, total biomass, recruitment, and fishing mortality time-series across models
}
\description{
Get Time Series Plots
}
\examples{
\dontrun{
  get_ts_plot(list(rep1, rep2), list(sd_rep1, sd_rep2), c("Model1", "Model2"))
}
}
