% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGA.R
\name{ladderCreate}
\alias{ladderCreate}
\title{ladderCreate}
\usage{
ladderCreate(laddername = "ladder.csv", repladder = 3)
}
\description{
Function to create ladder, the output of this function is a 3rd order polynomial,
 which is the best fit of the data. This function expects to find a CSV of name
 ladder.csv (or user defined name) in the current working directory, this should
 be a user generated file
}
