% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGA.R
\name{OGA}
\alias{OGA}
\title{OGA (Outgrowth Analysis)}
\usage{
OGA(
  autoinput = TRUE,
  nestdataframe = dfs,
  homedir = getwd(),
  ladder = TRUE,
  laddername = "ladder.csv",
  repladder = 3,
  blank = TRUE,
  blankname = "BLANK",
  create.plot = TRUE,
  stringencyfilt = 3,
  frequencyfilt = 0.05,
  measureInterval = 15,
  lowerlimitslope = 0.01,
  maxinflectionpoint = 1.3,
  LimitNoGrowth = 0.9,
  statsDT = TRUE,
  bacteria = 45,
  laglimit = 0.1,
  rmflagged = TRUE
)
}
\arguments{
\item{autoinput}{BOOL parameter, if TRUE then automatically read in all .csv files in current working directory for outgrowth
analysis. These csv files are autoinput into data frames, otherwise OGA expects to find a value for data frames.}

\item{nestdataframe}{if autoinput is FALSE, then instead you need to pass a nested dataframe to analyze
to this parameter.}

\item{homedir}{Home Directory= working directory containing RAW .csv files from plate reader, in Example.csv format.}

\item{ladder}{BOOL parameter, if TRUE then OGA creates calibration ladder from laddername and calibrates RAW data.}

\item{laddername}{Name of ladder.csv file (in the same directory as RAW .csv files).}

\item{repladder}{Number of technical replicates in ladder (see example).}

\item{blank}{Do you have blanks (wells with no inoculum and only media) in the csv files (TRUE/FALSE)}

\item{blankname}{Name used for blank wells, this name is case sensitive and must match exactly with the csv files.}

\item{create.plot}{BOOL parameter, if TRUE then OGA passes RAW files to create.plots to produce plots? (TRUE/FALSE).}

\item{stringencyfilt}{What order to use for stringency of Butterworth Filter.}

\item{frequencyfilt}{What cutoff to use for frequency of Butterworth Filter.}

\item{measureInterval}{Time between OD measurements in minutes.}

\item{lowerlimitslope}{Slope above which growth is considered exponential.}

\item{maxinflectionpoint}{OD above which the exponential growth phase cannot occur.}

\item{LimitNoGrowth}{What is the OD limit to signal no growth detected/insufficient information for complete growth curve.}

\item{statsDT}{BOOL parameter, if TRUE, OGA passes results files to statsDT(TRUE/FALSE)}

\item{bacteria}{Limit below which doubling-time is marked as bacterial contamination, for yeast. If measuring bacterial growth, set = 0.5.}

\item{laglimit}{OD above which marks the end of the lag-phase.}

\item{rmflagged}{This determines if flagged wells are removed from statistics. default=TRUE}
}
\description{
OGA calculates doubling times. The process is to subtract  blank wells, then smooth high frequency noise using a
Butterworth filter. OGA then identifies an interval of exponential growth. This is done by
picking an end time for exponential growth by defining the inflection point of exponential
growth as determined by the 1st derivative (diff) ul, and a beginning time for exponential growth by a heuristically
defined minimum slope change. The doubling time is then calculated between these begining and end points.
}
