% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGA.R
\name{OGA}
\alias{OGA}
\title{OGA (Outgrowth Analysis)}
\usage{
OGA(
  autoinput = TRUE,
  nestdataframe = dfs,
  homedir = getwd(),
  ladder = TRUE,
  laddername = "ladder.csv",
  repladder = 3,
  blank = TRUE,
  blankname = "BLANK",
  create.plot = TRUE,
  stringencyfilt = 3,
  frequencyfilt = 0.05,
  measureInterval = 15,
  lowerlimitslope = 0.01,
  maxinflectionpoint = 1.3,
  LimitNoGrowth = 0.9,
  stats = TRUE,
  bacteria = 45,
  laglimit = 0.1
)
}
\arguments{
\item{autoinput}{BOOL parameter, if TRUE then automatically read in all .csv files for outgrowth
analysis. these csv files are autoinputed into DFS, otherwise OGA expect to find a value for dfs.}

\item{nestdataframe}{if autoinput is FALSE then you need to pass instead a nested dataframe to analyze
to this paramater.}

\item{homedir}{Home Directory}

\item{ladder}{ladder}

\item{laddername}{Name of ladder}

\item{repladder}{Number of replicates in ladder}

\item{blank}{Do you have blanks in the csv files (TRUE/FALSE)}

\item{blankname}{Name used for blanks, this name is case sensitive and must match exactly with file.}

\item{create.plot}{Should we produce plots? (TRUE/FALSE)}

\item{stringencyfilt}{What filter to use for stringency.}

\item{frequencyfilt}{What filter to use for frequency.}

\item{measureInterval}{Time between measurements.}

\item{lowerlimitslope}{What is the lower limit for the slope to calculate.}

\item{maxinflectionpoint}{What is the maximum inflection point.}

\item{LimitNoGrowth}{What is the OD limit to signal no growth detected.}

\item{stats}{Do you want to run stats (TRUE/FALSE)}

\item{bacteria}{Should OGA flag potential bacterial growth.}

\item{laglimit}{Limit to how far until you declare a lag state.}
}
\description{
OGA calculates doubling times by subtracting blank wells, smoothing high frequency noise by a
Butterworth filter, picking the second limit by defining the inflection point of exponential
growth as determined by the 1st derivative (diff) ul, and first limit by a heuristically
defined slope change the doubling time is then calculated between these points.
}
\examples{
This first example assumes that your experiment measured OD values every 15 minutes and that you
labeled blanks with the keyword BLANK in the csv files containing the data. This is what the
default values are set to so just call the function.

> OGA()

This next example changes the time interval of the measurements to be every 30 minutes.

> OGA(measureInterval = 30)
}
