% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Biologicals.R
\name{Setup_Mod_Biologicals}
\alias{Setup_Mod_Biologicals}
\title{Setup biological inputs for estimation model}
\usage{
Setup_Mod_Biologicals(
  input_list,
  WAA,
  MatAA,
  AgeingError = NULL,
  Use_M_prior = 0,
  M_prior = NA,
  fit_lengths = 0,
  SizeAgeTrans = NA,
  M_spec = NULL,
  Fixed_natmort = NULL,
  ...
)
}
\arguments{
\item{input_list}{List containing a data list, parameter list, and map list}

\item{WAA}{Inputs of weight at age dimensioned by n_region, n_year, n_ages, n_sexes}

\item{MatAA}{Inputs of maturity at age dimensioned by n_region, n_year, n_ages, n_sexes}

\item{AgeingError}{Inputs of ageing error matrix dimensioned by number of modelled ages, number of observed composition ages Default behavior uses an identity matrix (no ageing error)}

\item{Use_M_prior}{Whether or not to use natural mortality prior, == 0 don't use, == 1 use}

\item{M_prior}{Vector of natural mortality priors with the first element representing the mean in normal space, and the second element representing the sd.}

\item{fit_lengths}{Whether or not to fit length data, == 0 dont fit, == 1 fit}

\item{SizeAgeTrans}{Size age transition matrix dimensioned by n_regions, n_years, n_lens, n_ages, n_sexes}

\item{M_spec}{Character specifying options for how to estimate natural mortality. Default is NULL such that it is estimated for 2 sexes or only estimated for a single sex if it is a single sex model. Other options include "est_shared_s" which estimates the same natural mortality rate if n_sexes == 2. The other option is "fix" which fixes all natural mortality parameters.}

\item{Fixed_natmort}{Fixed natural mortality array, dimensionsed by n_regions, n_yrs, n_ages, and n_sexes}

\item{...}{Additional arguments for starting values of ln_M and M_offset (note that when M_spec is specified at 'fix', these will not be used. Instead, a user must supply a fixed natural mortality array using Fixed_natmort)}
}
\description{
Setup biological inputs for estimation model
}
