% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Dimensions.R
\name{Setup_Mod_Dim}
\alias{Setup_Mod_Dim}
\title{Set up model dimensions}
\usage{
Setup_Mod_Dim(
  years,
  ages,
  lens,
  n_regions,
  n_sexes,
  n_fish_fleets,
  n_srv_fleets
)
}
\arguments{
\item{years}{vector of years}

\item{ages}{vector of ages}

\item{lens}{vector of lengths (can just input 1 if not fitting lengths)}

\item{n_regions}{Number of regions}

\item{n_sexes}{Number of sexes}

\item{n_fish_fleets}{Number of fishery fleets}

\item{n_srv_fleets}{Number of survey fleets}
}
\value{
Returns a list object of length 3, with a data list, a par list, and a map list
}
\description{
Set up model dimensions
}
