% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Fishery.R
\name{Setup_Mod_FishIdx_and_Comps}
\alias{Setup_Mod_FishIdx_and_Comps}
\title{Setup observed fishery indices and composition data (age and length comps)}
\usage{
Setup_Mod_FishIdx_and_Comps(
  input_list,
  ObsFishIdx,
  ObsFishIdx_SE,
  fish_idx_type,
  UseFishIdx,
  ObsFishAgeComps,
  UseFishAgeComps,
  ISS_FishAgeComps = NULL,
  ObsFishLenComps,
  UseFishLenComps,
  ISS_FishLenComps = NULL,
  FishAgeComps_LikeType,
  FishLenComps_LikeType,
  FishAgeComps_Type,
  FishLenComps_Type,
  FishAge_comp_agg_type = NULL,
  FishLen_comp_agg_type = NULL,
  ...
)
}
\arguments{
\item{input_list}{List containing a data list, parameter list, and map list}

\item{ObsFishIdx}{Observed fishery indices as an array dimensioned by n_regions, n_years, n_fish_fleets}

\item{ObsFishIdx_SE}{Observed standard errors for fishery indices as an array dimensioned by n_regions, n_years, n_fish_fleets}

\item{fish_idx_type}{Fishery index type dimensioned by n_fish_fleets, character string for abundance (abd) or biomass (biom), "none" if not availabile or used}

\item{UseFishIdx}{Indicator variables specifying whether or not to fit fishery indices as an array imensioned by n_regions, n_years, n_fish_fleets, == 0 don't fit, == 1 fit}

\item{ObsFishAgeComps}{Observed fishery age compositions as an array dimensioned by n_regions, n_years, n_ages, n_sexes, n_fish_fleets (should be in numbers, although need not be whole numbers)}

\item{UseFishAgeComps}{Indicator variables specifying whether or not to fit fishery ages as an array imensioned by n_regions, n_years, n_fish_fleets, == 0 don't fit, == 1 fit}

\item{ISS_FishAgeComps}{Input sample size for fishery comps dimensioned by n_regions, n_years, n_sexes, and n_fleets. If your sample sizes are already defined in your observed comps (i.e., numbers that don't sum to 1, can leave as is, but if they do sum to 1, you need to be sure to define this ISS so comps are correctly weighted (not given a ISS weight of 1))}

\item{ObsFishLenComps}{Observed fishery age compositions as an array dimensioned by n_regions, n_years, n_lens, n_sexes, n_fish_fleets (should be in numbers, although need not be whole numbers)}

\item{UseFishLenComps}{Indicator variables specifying whether or not to fit fishery lengths as an array imensioned by n_regions, n_years, n_fish_fleets, == 0 don't fit, == 1 fit}

\item{ISS_FishLenComps}{Input sample size for fishery comps dimensioned by n_regions, n_years, n_sexes, and n_fleets. If your sample sizes are already defined in your observed comps (i.e., numbers that don't sum to 1, can leave as is, but if they do sum to 1, you need to be sure to define this ISS so comps are correctly weighted (not given a ISS weight of 1))}

\item{FishAgeComps_LikeType}{Vector of character strings dimensioned by n_fish_fleets, options include Multinomial, Dirichlet-Multinomial, and iid-Logistic-Normal. Can specify with "none" if no likelihoods are used.}

\item{FishLenComps_LikeType}{Vector of character strings dimensioned by n_fish_fleets, options include Multinomial, Dirichlet-Multinomial, and iid-Logistic-Normal. Can specify with "none" if no likelihoods are used.}

\item{FishAgeComps_Type}{Character vector defines the compositiont and that is structured as: the composition type (agg, spltRspltS, spltRjntS), _, Year, _, year range you want the composition type to be in, _, Fleet, _, Fleet number. e.g., ("agg_Year_1-10_Fleet_1", "spltRjntS_Year_11-terminal_Fleet_1") species compositions to be aggregated in years 1 - 10 for fleet 1, split by region and sex in years 11 - terminal for fleet 1.}

\item{FishLenComps_Type}{Character vector defines the compositiont and that is structured as: the composition type (agg, spltRspltS, spltRjntS, none), _, Year, _, year range you want the composition type to be in, _, Fleet, _, Fleet number. e.g., ("agg_Year_1-10_Fleet_1", "spltRjntS_Year_11-terminal_Fleet_1") species compositions to be aggregated in years 1 - 10 for fleet 1, split by region and sex in years 11 - terminal for fleet 1.}

\item{FishAge_comp_agg_type}{Vector dimensioned by n_fish_fleets specifying how to aggregate age composition data if FishAgeComps_Type == 0, if comp_agg_type == 0 normalize, aggregate, ageing erorr then normalize, == 1 aggregate, normalize, then ageing error (for age compositions). Default is specify at NULL}

\item{FishLen_comp_agg_type}{Vector dimensioned by n_fish_fleets specifying how to aggregate age composition data if FishLenComps_Type == 1, if comp_agg_type == 0 length compositions are not normalized prior to application of size age transition, == 1 length compositions are normalized and then a size-age transition is applied (for length compositions). Default is specify at NULL}

\item{...}{Additional arguments specifying starting values for survey age and length overdispersion parameters (ln_FishAge_theta, ln_FishLen_theta, ln_FishAge_theta_agg, ln_FishLen_theta_agg)}
}
\description{
Setup observed fishery indices and composition data (age and length comps)
}
