% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Survey.R
\name{Setup_Mod_SrvIdx_and_Comps}
\alias{Setup_Mod_SrvIdx_and_Comps}
\title{Setup observed survey indices and composition data (age and length comps)}
\usage{
Setup_Mod_SrvIdx_and_Comps(
  input_list,
  ObsSrvIdx,
  ObsSrvIdx_SE,
  UseSrvIdx,
  srv_idx_type,
  ObsSrvAgeComps,
  UseSrvAgeComps,
  ObsSrvLenComps,
  UseSrvLenComps,
  ISS_SrvAgeComps = NULL,
  ISS_SrvLenComps = NULL,
  SrvAgeComps_LikeType,
  SrvLenComps_LikeType,
  SrvAgeComps_Type,
  SrvLenComps_Type,
  SrvAge_comp_agg_type = NULL,
  SrvLen_comp_agg_type = NULL,
  ...
)
}
\arguments{
\item{input_list}{List containing a data list, parameter list, and map list}

\item{ObsSrvIdx}{Observed survey indices as an array dimensioned by n_regions, n_years, n_srv_fleets}

\item{ObsSrvIdx_SE}{Observed standard errors for survey indices as an array dimensioned by n_regions, n_years, n_srv_fleets}

\item{UseSrvIdx}{Indicator variables specifying whether or not to fit survey indices as an array imensioned by n_regions, n_years, n_srv_fleets, == 0 don't fit, == 1 fit}

\item{srv_idx_type}{Survey index type dimensioned by n_srv_fleets, character string for abundance (abd) or biomass (biom), "none" if not availabile or used}

\item{ObsSrvAgeComps}{Observed survey age compositions as an array dimensioned by n_regions, n_years, number of observed composition ages, n_sexes, n_srv_fleets (should be in numbers, although need not be whole numbers)}

\item{UseSrvAgeComps}{Indicator variables specifying whether or not to fit survey ages as an array imensioned by n_regions, n_years, n_srv_fleets, == 0 don't fit, == 1 fit}

\item{ObsSrvLenComps}{Observed survey age compositions as an array dimensioned by n_regions, n_years, n_lens, n_sexes, n_srv_fleets (should be in numbers, although need not be whole numbers)}

\item{UseSrvLenComps}{Indicator variables specifying whether or not to fit survey lengths as an array imensioned by n_regions, n_years, n_srv_fleets, == 0 don't fit, == 1 fit}

\item{ISS_SrvAgeComps}{Input sample size for survey comps. Dimensioned by n_regions, n_years, n_sexes, n_fleets. Can be left as is if numbers are defined for the observed comps because the function sums them up interally to get an ISS. However, if observed comps are input as proportions, be sure to specify an ISS so the model doesn't incorrectly weight it as a value of ISS = 1.}

\item{ISS_SrvLenComps}{Input sample size for survey comps. Dimensioned by n_regions, n_years, n_sexes, n_fleets. Can be left as is if numbers are defined for the observed comps because the function sums them up interally to get an ISS. However, if observed comps are input as proportions, be sure to specify an ISS so the model doesn't incorrectly weight it as a value of ISS = 1.}

\item{SrvAgeComps_LikeType}{Vector of character strings dimensioned by n_srv_fleets, options include Multinomial, Dirichlet-Multinomial, and iid-Logistic-Normal. Can specify with "none" if no likelihoods are used.}

\item{SrvLenComps_LikeType}{Vector of character strings dimensioned by n_srv_fleets, options include Multinomial, Dirichlet-Multinomial, and iid-Logistic-Normal. Can specify with "none" if no likelihoods are used.}

\item{SrvAgeComps_Type}{Defines how age composition data are fit to by sex and region, with four possible options (agg, spltRspltS, spltRjntS, none). Option agg indicates that compositions are aggregated across regions and sexes, option spltRspltS indicates that compositions are split by region and split by sex (compositions sum to 1 within a given region and sex), option spltRjntS indicates that compositions are split by region but sum to 1 jointly across sexes, and option none specifies that no composition data are available. After defining the composition type, users will need to specify the year range you want the composition type to be in, as well as the fleet number the composition type belongs to. For example, ("agg_Year_1-10_Fleet_1", "spltRjntS_Year_11-terminal_Fleet_1") specifies compositions to be aggregated in years 1 - 10 for fleet 1, split by region and sex in years 11 - terminal for fleet 1.}

\item{SrvLenComps_Type}{Defines how length composition data are fit to by sex and region, with four possible options (agg, spltRspltS, spltRjntS, none). Option agg indicates that compositions are aggregated across regions and sexes, option spltRspltS indicates that compositions are split by region and split by sex (compositions sum to 1 within a given region and sex), option spltRjntS indicates that compositions are split by region but sum to 1 jointly across sexes, and option none specifies that no composition data are available. After defining the composition type, users will need to specify the year range you want the composition type to be in, as well as the fleet number the composition type belongs to. For example, ("agg_Year_1-10_Fleet_1", "spltRjntS_Year_11-terminal_Fleet_1") specifies compositions to be aggregated in years 1 - 10 for fleet 1, split by region and sex in years 11 - terminal for fleet 1.}

\item{SrvAge_comp_agg_type}{Vector dimensioned by n_srv_fleets specifying how to aggregate age composition data if SrvAgeComps_Type == 0, if comp_agg_type == 0 normalize, aggregate, ageing erorr then normalize, == 1 aggregate, normalize, then ageing error (for age compositions). Default is specify at NULL}

\item{SrvLen_comp_agg_type}{Vector dimensioned by n_srv_fleets specifying how to aggregate age composition data if SrvLenComps_Type == 1, if comp_agg_type == 0 length compositions are not normalized prior to application of size age transition, == 1 length compositions are normalized and then a size-age transition is applied (for length compositions). Default is specify at NULL}

\item{...}{Additional arguments specifying starting values for survey age and length overdispersion parameters (ln_SrvAge_theta. ln_SrvLen_theta, ln_SrvAge_theta_agg, ln_SrvLen_theta_agg)}
}
\description{
Setup observed survey indices and composition data (age and length comps)
}
