% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Tagging.R
\name{Setup_Mod_Tagging}
\alias{Setup_Mod_Tagging}
\title{Setup tagging processes and parameters}
\usage{
Setup_Mod_Tagging(
  input_list,
  UseTagging = 0,
  tag_release_indicator = NULL,
  max_tag_liberty = 0,
  Tagged_Fish = NA,
  Obs_Tag_Recap = NA,
  Tag_LikeType = NA,
  mixing_period = 1,
  t_tagging = 0,
  tag_selex = NA,
  tag_natmort = NA,
  Use_TagRep_Prior = 0,
  TagRep_PriorType = NA,
  TagRep_mu = NA,
  TagRep_sd = NA,
  move_age_tag_pool = NA,
  move_sex_tag_pool = NA,
  Init_Tag_Mort_spec = NULL,
  Tag_Shed_spec = NULL,
  TagRep_spec = NULL,
  Tag_Reporting_blocks = NULL,
  ...
)
}
\arguments{
\item{input_list}{List containing a data list, parameter list, and map list}

\item{UseTagging}{Numeric indicating whether to use tagging data (1) or not (0)}

\item{tag_release_indicator}{Matrix dimensioned by n_tag_cohorts, 2 (where the first column is the release region, and the second column is the release year) describing the release region and year of a given tag cohort}

\item{max_tag_liberty}{Maximum number of years to track a tagged cohort}

\item{Tagged_Fish}{Array dimensioned by n_tag_cohorts, n_ages, n_sexes describing the tag released fish}

\item{Obs_Tag_Recap}{Observed tag recaptures dimensioned by max_tag_liberty, n_tag_cohorts, n_regions, n_ages, n_sexes}

\item{Tag_LikeType}{Numeric indicating tag likelihood type, == 0 Poisson, == 1 Negative Binomial, == 2 Multinomial Release Conditioned, == 3 Multinomial recapture conditioned}

\item{mixing_period}{Numeric indicating mixing period such that any year < mixing period for a given tag cohort is not fit to}

\item{t_tagging}{Fractional year in which tag releases happen}

\item{tag_selex}{Numeric indicating how tag recovery selectivity is parameterized, == 0 uniform selectivity, == 1 sex-averaged fishery selectivity from dominant fleet (fleet 1), == 2 sex-specific selectivity from dominant fleet (fleet 1)}

\item{tag_natmort}{Numeric indicating how tag natural mortality is parameterized, == 0 age- and sex-averaged, == 1 unique age, but sex-averaged natural mortality, == 2 unique sex, but age-aggregated natural mortality, == 3 unique sex-and age natural mortality}

\item{Use_TagRep_Prior}{Numeric indicating whether to use tag reporting rate prior == 0, don't use, == 1 use}

\item{TagRep_PriorType}{Numeric indicating the type of tag reporting prior to use, == 0 Symmetric beta, such that extremes are penalized heaviest, == 1 Regular Beta}

\item{TagRep_mu}{Numeric value indicating the mean of the tag reporting rate prior (in normal space), can be specified as NA if using a symmetric beta}

\item{TagRep_sd}{Numeric value indicating the sd of the tag reporting rate prior (in normal space), for a symmetric beta, smaller values impose a larger penalty along the edges}

\item{move_age_tag_pool}{List indicating how to fit tagging age-specific data. For example, if no age-specific data are avalialble, then the list should be dimensioned as lenght 1, where the first elmeent is 1:n_ages, however, if we have ageing data and we want to pool to fit data, then each element is the ages we want to pool. If we want to fit each age uniquely, then the list will have unique elements representing each age e.g., list(1,2,3,4, ... n_ages)}

\item{move_sex_tag_pool}{List indicating how to fit tagging sex-specific data. For example, if no sex-specific data are avalialble, then the list should be dimensioned as lenght 1, where the first elmeent is 1:n_sexes. However, if we have sex-specific data, this can be dimensioned as a length of 2 if we want to fit sex-specific data (e.g., list(1, 2))}

\item{Init_Tag_Mort_spec}{Specificaiotn of initial tag mortality. Options include fix, which fixes this parameter, or est, which estiamtes this parameter}

\item{Tag_Shed_spec}{Specificaiotn of chronic tag shedding. Options include fix, which fixes this parameter, or est, which estiamtes this parameter}

\item{TagRep_spec}{Specification of tag reporting rates. Options include est_all, which estiamtes all regions and blocks specified, est_shared_r, which estamites all blocks, but shares across regions.}

\item{Tag_Reporting_blocks}{Tag reporting blocks we want to specify. Character vector for each region "Block_1_Year_1-15_Region_1", "Block_2_Year_16-30_Region_1", "none_Region_2 to specify as an example.}

\item{...}{Additional arguments specifying starting values for tagging parameters (ln_Init_Tag_Mort, ln_Tag_Shed, ln_tag_theta, Tag_Reporting_Pars)}
}
\description{
Setup tagging processes and parameters
}
