% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Observation_Processes.R
\name{Setup_Mod_Weighting}
\alias{Setup_Mod_Weighting}
\title{Set up SPoCK model weighting}
\usage{
Setup_Mod_Weighting(
  input_list,
  sablefish_ADMB,
  likelihoods,
  Wt_Catch = 1,
  Wt_FishIdx = 1,
  Wt_SrvIdx = 1,
  Wt_Rec = 1,
  Wt_F = 1,
  Wt_Tagging = 1,
  Wt_FishAgeComps,
  Wt_SrvAgeComps,
  Wt_FishLenComps,
  Wt_SrvLenComps
)
}
\arguments{
\item{input_list}{List containing data, parameter, and map lists.}

\item{sablefish_ADMB}{Numeric flag to mimic calculations for the sablefish ADMB model:
\itemize{
  \item 1 to mimic sablefish ADMB calculations
  \item 0 otherwise
}}

\item{likelihoods}{Numeric flag indicating likelihood implementation to use:
\itemize{
  \item 0 for ADMB likelihoods
  \item 1 for TMB likelihoods
}}

\item{Wt_Catch}{Either a numeric scalar (lambda) applied to the overall catch dataset or an array of lambdas (i.e., weights can change by year and fleet) dimensioned by n_regions, n_years, n_fish_fleets.}

\item{Wt_FishIdx}{Either a numeric scalar (lambda) applied to the overall fishery index dataset  or an array of lambdas (i.e., weights can change by year and fleet) dimensioned by n_regions, n_years, n_fish_fleets.}

\item{Wt_SrvIdx}{Either a numeric scalar (lambda) applied to the overall survey index dataset or an array of lambdas (i.e., weights can change by year and fleet) dimensioned by n_regions, n_years, n_srv_fleets.}

\item{Wt_Rec}{Numeric weight (lambda) applied to the recruitment penalty.}

\item{Wt_F}{Numeric weight (lambda) applied to the fishing mortality penalty.}

\item{Wt_Tagging}{Numeric weight (lambda) applied to tagging data.}

\item{Wt_FishAgeComps}{Numeric weight (lambda) applied to fishery age composition data.}

\item{Wt_SrvAgeComps}{Numeric weight (lambda) applied to survey age composition data.}

\item{Wt_FishLenComps}{Numeric weight (lambda) applied to fishery length composition data.}

\item{Wt_SrvLenComps}{Numeric weight (lambda) applied to survey length composition data.}
}
\description{
Set up SPoCK model weighting
}
