% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Biologicals.R
\name{Setup_Sim_Biologicals}
\alias{Setup_Sim_Biologicals}
\title{Set up simulation stuff for biologicals}
\usage{
Setup_Sim_Biologicals(
  base_M_value,
  M_pattern,
  base_WAA_values,
  base_WAA_fish_values,
  WAA_pattern,
  base_Maturity_AA_values,
  Maturity_AA_pattern,
  sim_list
)
}
\arguments{
\item{base_M_value}{Base natural mortality value dimensioned by regions, ages, sexes}

\item{M_pattern}{Natural mortality pattern. Options include: constant}

\item{base_WAA_values}{Base weight-at-age values dimensioned by regions, ages, sexes}

\item{base_WAA_fish_values}{Base weight-at-age values for the fishery dimensioned by regions, ages, sexes, fishery fleets}

\item{WAA_pattern}{Weight-at-age pattern. Options include: constant}

\item{base_Maturity_AA_values}{Base maturity values dimensioned by regions, ages, sexes}

\item{Maturity_AA_pattern}{Maturity pattern. Options include: constant}

\item{sim_list}{Simulation list objects}
}
\description{
Set up simulation stuff for biologicals
}
