% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Containers.R
\name{Setup_Sim_Containers}
\alias{Setup_Sim_Containers}
\title{Setup containers for simulation and output into global environment}
\usage{
Setup_Sim_Containers(sim_list)
}
\arguments{
\item{sim_list}{List object containing simulation components (should have n_regions, n_ages, n_sexes, n_sims, sim_list$n_fish_fleets, sim_list$n_srv_fleets)}
}
\description{
Setup containers for simulation and output into global environment
}
