% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Fishery.R
\name{Setup_Sim_FishSel}
\alias{Setup_Sim_FishSel}
\title{Setup fishery selectivity}
\usage{
Setup_Sim_FishSel(sim_list, sel_model, fixed_fish_sel_pars)
}
\arguments{
\item{sim_list}{Simulation list}

\item{sel_model}{Fishery selectivity model dimensioned by region and fleet. Options include: logistic}

\item{fixed_fish_sel_pars}{Fixed parameters of fishery selectivity, dimensioned by region, sex, fishery fleet, and the max number of parameters needed for
for a defined fishery selectivity functional form out of all defined functional forms for the fishery}
}
\description{
Setup fishery selectivity
}
