% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Rec.R
\name{Setup_Sim_Rec}
\alias{Setup_Sim_Rec}
\title{Set up recruitment dynamics for simulation}
\usage{
Setup_Sim_Rec(
  n_yrs,
  n_regions,
  n_sexes,
  n_sims,
  do_recruits_move,
  base_rec_sexratio,
  rec_sexratio_vary,
  base_r0,
  r0_vary,
  base_h,
  init_sigmaR,
  sigmaR,
  recruitment_opt,
  recdev_opt
)
}
\arguments{
\item{n_yrs}{Number of years}

\item{n_regions}{Number of regions}

\item{n_sexes}{Number of sexes}

\item{n_sims}{Number of simulations}

\item{do_recruits_move}{whether recruits move == 0, don't move, == 1 move}

\item{base_rec_sexratio}{base recruitment sex-ratio value}

\item{rec_sexratio_vary}{whether recruitment sex-ratio varies. Options include: constant}

\item{base_r0}{base R0 or mean recruitment value}

\item{r0_vary}{whether r0 or mean recruitment varies. Options include: constant}

\item{base_h}{base steepness value}

\item{init_sigmaR}{Sigma R for initial devs}

\item{sigmaR}{Sigma R for everything else}

\item{recruitment_opt}{== 0, mean recruitment, == 1 Beverton Holt}

\item{recdev_opt}{== 0, global rec devs, == 1, local rec devs}
}
\description{
Set up recruitment dynamics for simulation
}
