% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Survey.R
\name{Setup_Sim_Survey}
\alias{Setup_Sim_Survey}
\title{Setup values for survey parameterization}
\usage{
Setup_Sim_Survey(
  sigmaSrvIdx,
  base_srv_q,
  srv_q_pattern,
  sel_model,
  fixed_srv_sel_pars,
  sim_list
)
}
\arguments{
\item{sigmaSrvIdx}{Survey index observation error, dimensioned by region and fleet}

\item{base_srv_q}{Base survey catchability value, dimensioned by region and fleet}

\item{srv_q_pattern}{Survey catchability pattern, dimensioned by region and fleet. Options include: constant}

\item{sel_model}{Survey selectivity model dimensioned by region and fleet. Options include: logistic}

\item{fixed_srv_sel_pars}{Fixed parameters of survey selectivity, dimensioned by region, sex, survey fleet, and the max number of parameters needed for
for a defined survey selectivity functional form out of all defined functional forms for the survey.}

\item{sim_list}{Simulation list object}
}
\description{
Setup values for survey parameterization
}
