% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Tagging.R
\name{Setup_Sim_Tagging}
\alias{Setup_Sim_Tagging}
\title{Set up simulated tagging dynamics}
\usage{
Setup_Sim_Tagging(
  n_tags,
  max_liberty,
  tag_years,
  t_tagging,
  base_Tag_Reporting,
  Tag_Reporting_pattern,
  Tag_Ind_Mort,
  Tag_Shed,
  sim_list
)
}
\arguments{
\item{n_tags}{Number of tags to release in a given year}

\item{max_liberty}{Maximum liberty to track cohorts for}

\item{tag_years}{Years to release tags for}

\item{t_tagging}{Time of tagging (e.g., start year == 0, mid year == 0.5)}

\item{base_Tag_Reporting}{Base tag reporting rate by region}

\item{Tag_Reporting_pattern}{Tag reporting pattern. Options include: constant}

\item{Tag_Ind_Mort}{Initial tag induced mortality}

\item{Tag_Shed}{Chronic tag shedding rate}

\item{sim_list}{Simulation list}
}
\description{
Set up simulated tagging dynamics
}
