% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Reference_Points.R
\name{single_region_SPR}
\alias{single_region_SPR}
\title{Get SPR reference points}
\usage{
single_region_SPR(pars, data)
}
\arguments{
\item{pars}{Parameter List}

\item{data}{Data List}
}
\description{
Get SPR reference points
}
\examples{
\dontrun{
# extract out elements from report
F_fract_flt <- rep$Fmort[1,65,] / sum(rep$Fmort[1,65,])
fish_sel <- rep$fish_sel[1,65,,1,]
natmort <- rep$natmort[1,65,,1]
t_spwn <- 0
WAA <- rtmb_data$WAA[1,65,,1]
MatAA <- rtmb_data$MatAA[1,65,,1]
SPR_x <- 0.4 # SPR \% to target
# set up data list
data <- list(F_fract_flt = F_fract_flt,
             fish_sel = fish_sel,
             natmort = natmort,
             t_spwn = t_spwn,
             WAA = WAA,
             MatAA = MatAA,
             SPR_x = SPR_x)
pars <- list(log_F_x = 0.4)
# make adfun and run model
spr_ad <- RTMB::MakeADFun(cmb(single_region_SPR, data), parameters = pars, map = NULL)
optim <- stats::nlminb(spr_ad$par, spr_ad$fn, spr_ad$gr,
                       control = list(iter.max = 1e5, eval.max = 1e5, rel.tol = 1e-15))
exp(optim$par) # f_x
}
}
