% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{statsCLS}
\alias{statsCLS}
\title{stats}
\usage{
statsCLS()
}
\description{
stats takes raw results files and calculates the mean, SD and SEM of all wells with the same name from SurvivalPercentage and SurvivalIntegral.
}
