% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.R
\name{statsDT}
\alias{statsDT}
\title{statsDT}
\usage{
statsDT(locationofresults = "Results", LimitNoGrowth = 0.9, rmflagged = TRUE)
}
\arguments{
\item{locationofresults}{path to the results of OGA analysis.}

\item{LimitNoGrowth}{Parameter specifying the cutoff value for determining growth. default=0.9}

\item{rmflagged}{This determines if flagged wells are removed from statistics. default=TRUE}
}
\description{
statsDT takes raw results files and calculates the mean, SD and SEM of all wells with the same name.
}
