% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLCalc.R
\name{SurvivalCalc}
\alias{SurvivalCalc}
\title{SurvivalCalc}
\usage{
SurvivalCalc(
  firstDay = 1,
  resultspath = getwd(),
  rmflagged = TRUE,
  statsCLS = TRUE,
  measureInterval = 15
)
}
\arguments{
\item{firstDay}{Specify the first day of measurements, default=1}

\item{resultspath}{System path where the results are stored, default=current directory}

\item{rmflagged}{This determines if flagged wells are removed from CLS statistics, default=TRUE}

\item{statsCLS}{Should the code run statsCLS on the CLS results (TRUE/FALSE), default=TRUE}

\item{measureInterval}{Interval in minutes for each OD measurement, default=15}
}
\description{
SurvivalCalc removes nogrowth, contaminates and unexpected growth from experiments and then passes
the modified results to both survivalIntegral and survivalPercentage for calculations, followed by statsCLS for statistics.
}
